Network.AWS.CloudWatchLogs

Description

You can use Amazon CloudWatch Logs to monitor, store, and access your log files from EC2 instances, Amazon CloudTrail, or other sources. You can then retrieve the associated log data from CloudWatch Logs using the Amazon CloudWatch console, the CloudWatch Logs commands in the AWS CLI, the CloudWatch Logs API, or the CloudWatch Logs SDK.

You can use CloudWatch Logs to:

Monitor Logs from Amazon EC2 Instances in Real-time : You can use CloudWatch Logs to monitor applications and systems using log data. For example, CloudWatch Logs can track the number of errors that occur in your application logs and send you a notification whenever the rate of errors exceeds a threshold you specify. CloudWatch Logs uses your log data for monitoring; so, no code changes are required. For example, you can monitor application logs for specific literal terms (such as NullReferenceException) or count the number of occurrences of a literal term at a particular position in log data (such as "404" status codes in an Apache access log). When the term you are searching for is found, CloudWatch Logs reports the data to a Amazon CloudWatch metric that you specify.

Monitor Amazon CloudTrail Logged Events : You can create alarms in Amazon CloudWatch and receive notifications of particular API activity as captured by CloudTrail and use the notification to perform troubleshooting.

Archive Log Data : You can use CloudWatch Logs to store your log data in highly durable storage. You can change the log retention setting so that any log events older than this setting are automatically deleted. The CloudWatch Logs agent makes it easy to quickly send both rotated and non-rotated log data off of a host and into the log service. You can then access the raw log data when you need it.

Service Configuration

cloudWatchLogs :: Service

API version 2014-03-28 of the Amazon CloudWatch Logs SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CloudWatchLogs.

InvalidParameterException

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

A parameter is specified incorrectly.

InvalidSequenceTokenException

_InvalidSequenceTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The sequence token is not valid.

ResourceAlreadyExistsException

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource already exists.

OperationAbortedException

_OperationAbortedException :: AsError a => Getting (First ServiceError) a ServiceError

Multiple requests to update the same resource were in conflict.

ServiceUnavailableException

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service cannot complete the request.

DataAlreadyAcceptedException

_DataAlreadyAcceptedException :: AsError a => Getting (First ServiceError) a ServiceError

The event was already logged.

InvalidOperationException

_InvalidOperationException :: AsError a => Getting (First ServiceError) a ServiceError

The operation is not valid on the specified resource.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have reached the maximum number of resources that can be created.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DescribeDestinations (Paginated)

module Network.AWS.CloudWatchLogs.DescribeDestinations

UntagLogGroup

module Network.AWS.CloudWatchLogs.UntagLogGroup

CreateExportTask

module Network.AWS.CloudWatchLogs.CreateExportTask

PutDestination

module Network.AWS.CloudWatchLogs.PutDestination

DescribeSubscriptionFilters (Paginated)

module Network.AWS.CloudWatchLogs.DescribeSubscriptionFilters

GetLogEvents

module Network.AWS.CloudWatchLogs.GetLogEvents

DescribeLogGroups (Paginated)

module Network.AWS.CloudWatchLogs.DescribeLogGroups

DeleteDestination

module Network.AWS.CloudWatchLogs.DeleteDestination

FilterLogEvents (Paginated)

module Network.AWS.CloudWatchLogs.FilterLogEvents

TagLogGroup

module Network.AWS.CloudWatchLogs.TagLogGroup

DeleteLogStream

module Network.AWS.CloudWatchLogs.DeleteLogStream

CreateLogStream

module Network.AWS.CloudWatchLogs.CreateLogStream

CreateLogGroup

module Network.AWS.CloudWatchLogs.CreateLogGroup

DescribeExportTasks

module Network.AWS.CloudWatchLogs.DescribeExportTasks

CancelExportTask

module Network.AWS.CloudWatchLogs.CancelExportTask

PutSubscriptionFilter

module Network.AWS.CloudWatchLogs.PutSubscriptionFilter

DeleteLogGroup

module Network.AWS.CloudWatchLogs.DeleteLogGroup

DeleteSubscriptionFilter

module Network.AWS.CloudWatchLogs.DeleteSubscriptionFilter

PutLogEvents

module Network.AWS.CloudWatchLogs.PutLogEvents

DescribeMetricFilters (Paginated)

module Network.AWS.CloudWatchLogs.DescribeMetricFilters

TestMetricFilter

module Network.AWS.CloudWatchLogs.TestMetricFilter

PutDestinationPolicy

module Network.AWS.CloudWatchLogs.PutDestinationPolicy

PutMetricFilter

module Network.AWS.CloudWatchLogs.PutMetricFilter

DeleteRetentionPolicy

module Network.AWS.CloudWatchLogs.DeleteRetentionPolicy

DeleteMetricFilter

module Network.AWS.CloudWatchLogs.DeleteMetricFilter

PutRetentionPolicy

module Network.AWS.CloudWatchLogs.PutRetentionPolicy

ListTagsLogGroup

module Network.AWS.CloudWatchLogs.ListTagsLogGroup

DescribeLogStreams (Paginated)

module Network.AWS.CloudWatchLogs.DescribeLogStreams

Types

Distribution

data Distribution

Constructors

ByLogStream

Random

ExportTaskStatusCode

data ExportTaskStatusCode

Constructors

Cancelled

Completed

Failed

Pending

PendingCancel

Running

OrderBy

data OrderBy

Constructors

LastEventTime

LogStreamName

Destination

data Destination

Represents a cross-account destination that receives subscription log events.

See: destination smart constructor.

destination :: Destination

Creates a value of Destination with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dTargetARN - The Amazon Resource Name (ARN) of the physical target where the log events will be delivered (for example, a Kinesis stream).

dCreationTime - The creation time of the destination, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

dArn - The ARN of this destination.

dAccessPolicy - An IAM policy document that governs which AWS accounts can create subscription filters against this destination.

dDestinationName - The name of the destination.

dRoleARN - A role for impersonation, used when delivering log events to the target.

dTargetARN :: Lens' Destination (Maybe Text)

The Amazon Resource Name (ARN) of the physical target where the log events will be delivered (for example, a Kinesis stream).

dCreationTime :: Lens' Destination (Maybe Natural)

The creation time of the destination, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

dArn :: Lens' Destination (Maybe Text)

The ARN of this destination.

dAccessPolicy :: Lens' Destination (Maybe Text)

An IAM policy document that governs which AWS accounts can create subscription filters against this destination.

dDestinationName :: Lens' Destination (Maybe Text)

The name of the destination.

dRoleARN :: Lens' Destination (Maybe Text)

A role for impersonation, used when delivering log events to the target.

ExportTask

data ExportTask

Represents an export task.

See: exportTask smart constructor.

exportTask :: ExportTask

Creates a value of ExportTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

etDestinationPrefix - The prefix that was used as the start of Amazon S3 key for every object exported.

etDestination - The name of Amazon S3 bucket to which the log data was exported.

etStatus - The status of the export task.

etTaskName - The name of the export task.

etTaskId - The ID of the export task.

etTo - The end time, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

etFrom - The start time, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp prior to this time are not exported.

etLogGroupName - The name of the log group from which logs data was exported.

etExecutionInfo - Execution info about the export task.

etDestinationPrefix :: Lens' ExportTask (Maybe Text)

The prefix that was used as the start of Amazon S3 key for every object exported.

etDestination :: Lens' ExportTask (Maybe Text)

The name of Amazon S3 bucket to which the log data was exported.

etStatus :: Lens' ExportTask (Maybe ExportTaskStatus)

The status of the export task.

etTaskName :: Lens' ExportTask (Maybe Text)

The name of the export task.

etTaskId :: Lens' ExportTask (Maybe Text)

The ID of the export task.

etTo :: Lens' ExportTask (Maybe Natural)

The end time, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

etFrom :: Lens' ExportTask (Maybe Natural)

The start time, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp prior to this time are not exported.

etLogGroupName :: Lens' ExportTask (Maybe Text)

The name of the log group from which logs data was exported.

etExecutionInfo :: Lens' ExportTask (Maybe ExportTaskExecutionInfo)

Execution info about the export task.

ExportTaskExecutionInfo

data ExportTaskExecutionInfo

Represents the status of an export task.

See: exportTaskExecutionInfo smart constructor.

exportTaskExecutionInfo :: ExportTaskExecutionInfo

Creates a value of ExportTaskExecutionInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eteiCreationTime - The creation time of the export task, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

eteiCompletionTime - The completion time of the export task, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

eteiCreationTime :: Lens' ExportTaskExecutionInfo (Maybe Natural)

The creation time of the export task, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

eteiCompletionTime :: Lens' ExportTaskExecutionInfo (Maybe Natural)

The completion time of the export task, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

ExportTaskStatus

data ExportTaskStatus

Represents the status of an export task.

See: exportTaskStatus smart constructor.

exportTaskStatus :: ExportTaskStatus

Creates a value of ExportTaskStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

etsCode - The status code of the export task.

etsMessage - The status message related to the status code.

etsCode :: Lens' ExportTaskStatus (Maybe ExportTaskStatusCode)

The status code of the export task.

etsMessage :: Lens' ExportTaskStatus (Maybe Text)

The status message related to the status code.

FilteredLogEvent

data FilteredLogEvent

Represents a matched event.

See: filteredLogEvent smart constructor.

filteredLogEvent :: FilteredLogEvent

Creates a value of FilteredLogEvent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fleIngestionTime - The time the event was ingested, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

fleLogStreamName - The name of the log stream this event belongs to.

fleMessage - The data contained in the log event.

fleTimestamp - The time the event occurred, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

fleEventId - The ID of the event.

fleIngestionTime :: Lens' FilteredLogEvent (Maybe Natural)

The time the event was ingested, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

fleLogStreamName :: Lens' FilteredLogEvent (Maybe Text)

The name of the log stream this event belongs to.

fleMessage :: Lens' FilteredLogEvent (Maybe Text)

The data contained in the log event.

fleTimestamp :: Lens' FilteredLogEvent (Maybe Natural)

The time the event occurred, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

fleEventId :: Lens' FilteredLogEvent (Maybe Text)

The ID of the event.

InputLogEvent

data InputLogEvent

Represents a log event, which is a record of activity that was recorded by the application or resource being monitored.

See: inputLogEvent smart constructor.

inputLogEvent

Arguments

:: Natural
ileTimestamp

-> Text
ileMessage

-> InputLogEvent

Creates a value of InputLogEvent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ileTimestamp - The time the event occurred, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

ileMessage - The raw event message.

ileTimestamp :: Lens' InputLogEvent Natural

The time the event occurred, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

ileMessage :: Lens' InputLogEvent Text

The raw event message.

LogGroup

data LogGroup

Represents a log group.

See: logGroup smart constructor.

logGroup :: LogGroup

Creates a value of LogGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lgCreationTime - The creation time of the log group, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

lgMetricFilterCount - The number of metric filters.

lgArn - The Amazon Resource Name (ARN) of the log group.

lgLogGroupName - The name of the log group.

lgRetentionInDays - Undocumented member.

lgStoredBytes - The number of bytes stored.

lgCreationTime :: Lens' LogGroup (Maybe Natural)

The creation time of the log group, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

lgMetricFilterCount :: Lens' LogGroup (Maybe Int)

The number of metric filters.

lgArn :: Lens' LogGroup (Maybe Text)

The Amazon Resource Name (ARN) of the log group.

lgLogGroupName :: Lens' LogGroup (Maybe Text)

The name of the log group.

lgRetentionInDays :: Lens' LogGroup (Maybe Int)

Undocumented member.

lgStoredBytes :: Lens' LogGroup (Maybe Natural)

The number of bytes stored.

LogStream

data LogStream

Represents a log stream, which is a sequence of log events from a single emitter of logs.

See: logStream smart constructor.

logStream :: LogStream

Creates a value of LogStream with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsCreationTime - The creation time of the stream, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

lsUploadSequenceToken - The sequence token.

lsArn - The Amazon Resource Name (ARN) of the log stream.

lsFirstEventTimestamp - The time of the first event, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

lsLogStreamName - The name of the log stream.

lsStoredBytes - The number of bytes stored.

lsLastIngestionTime - The ingestion time, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

lsLastEventTimestamp - the time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. lastEventTime updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but may take longer in some rare situations.

lsCreationTime :: Lens' LogStream (Maybe Natural)

The creation time of the stream, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

lsUploadSequenceToken :: Lens' LogStream (Maybe Text)

The sequence token.

lsArn :: Lens' LogStream (Maybe Text)

The Amazon Resource Name (ARN) of the log stream.

lsFirstEventTimestamp :: Lens' LogStream (Maybe Natural)

The time of the first event, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

lsLogStreamName :: Lens' LogStream (Maybe Text)

The name of the log stream.

lsStoredBytes :: Lens' LogStream (Maybe Natural)

The number of bytes stored.

lsLastIngestionTime :: Lens' LogStream (Maybe Natural)

The ingestion time, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

lsLastEventTimestamp :: Lens' LogStream (Maybe Natural)

the time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. lastEventTime updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but may take longer in some rare situations.

MetricFilter

data MetricFilter

Metric filters express how CloudWatch Logs would extract metric observations from ingested log events and transform them into metric data in a CloudWatch metric.

See: metricFilter smart constructor.

metricFilter :: MetricFilter

Creates a value of MetricFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mfCreationTime - The creation time of the metric filter, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

mfFilterName - The name of the metric filter.

mfLogGroupName - The name of the log group.

mfFilterPattern - Undocumented member.

mfMetricTransformations - The metric transformations.

mfCreationTime :: Lens' MetricFilter (Maybe Natural)

The creation time of the metric filter, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

mfFilterName :: Lens' MetricFilter (Maybe Text)

The name of the metric filter.

mfLogGroupName :: Lens' MetricFilter (Maybe Text)

The name of the log group.

mfFilterPattern :: Lens' MetricFilter (Maybe Text)

Undocumented member.

mfMetricTransformations :: Lens' MetricFilter (Maybe (NonEmpty MetricTransformation))

The metric transformations.

MetricFilterMatchRecord

data MetricFilterMatchRecord

Represents a matched event.

See: metricFilterMatchRecord smart constructor.

metricFilterMatchRecord :: MetricFilterMatchRecord

Creates a value of MetricFilterMatchRecord with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mfmrExtractedValues - The values extracted from the event data by the filter.

mfmrEventNumber - The event number.

mfmrEventMessage - The raw event data.

mfmrExtractedValues :: Lens' MetricFilterMatchRecord (HashMap Text Text)

The values extracted from the event data by the filter.

mfmrEventNumber :: Lens' MetricFilterMatchRecord (Maybe Integer)

The event number.

mfmrEventMessage :: Lens' MetricFilterMatchRecord (Maybe Text)

The raw event data.

MetricTransformation

data MetricTransformation

Indicates how to transform ingested log events into metric data in a CloudWatch metric.

See: metricTransformation smart constructor.

metricTransformation

Arguments

:: Text
mtMetricName

-> Text
mtMetricNamespace

-> Text
mtMetricValue

-> MetricTransformation

Creates a value of MetricTransformation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mtDefaultValue - (Optional) The value to emit when a filter pattern does not match a log event. This value can be null.

mtMetricName - The name of the CloudWatch metric.

mtMetricNamespace - The namespace of the CloudWatch metric.

mtMetricValue - The value to publish to the CloudWatch metric when a filter pattern matches a log event.

mtDefaultValue :: Lens' MetricTransformation (Maybe Double)

(Optional) The value to emit when a filter pattern does not match a log event. This value can be null.

mtMetricName :: Lens' MetricTransformation Text

The name of the CloudWatch metric.

mtMetricNamespace :: Lens' MetricTransformation Text

The namespace of the CloudWatch metric.

mtMetricValue :: Lens' MetricTransformation Text

The value to publish to the CloudWatch metric when a filter pattern matches a log event.

OutputLogEvent

data OutputLogEvent

Represents a log event.

See: outputLogEvent smart constructor.

outputLogEvent :: OutputLogEvent

Creates a value of OutputLogEvent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

oleIngestionTime - The time the event was ingested, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

oleMessage - The data contained in the log event.

oleTimestamp - The time the event occurred, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

oleIngestionTime :: Lens' OutputLogEvent (Maybe Natural)

The time the event was ingested, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

oleMessage :: Lens' OutputLogEvent (Maybe Text)

The data contained in the log event.

oleTimestamp :: Lens' OutputLogEvent (Maybe Natural)

The time the event occurred, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

RejectedLogEventsInfo

data RejectedLogEventsInfo

Represents the rejected events.

See: rejectedLogEventsInfo smart constructor.

rejectedLogEventsInfo :: RejectedLogEventsInfo

Creates a value of RejectedLogEventsInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rleiTooOldLogEventEndIndex - The log events that are too old.

rleiTooNewLogEventStartIndex - The log events that are too new.

rleiExpiredLogEventEndIndex - The expired log events.

rleiTooOldLogEventEndIndex :: Lens' RejectedLogEventsInfo (Maybe Int)

The log events that are too old.

rleiTooNewLogEventStartIndex :: Lens' RejectedLogEventsInfo (Maybe Int)

The log events that are too new.

rleiExpiredLogEventEndIndex :: Lens' RejectedLogEventsInfo (Maybe Int)

The expired log events.

SearchedLogStream

data SearchedLogStream

Represents the search status of a log stream.

See: searchedLogStream smart constructor.

searchedLogStream :: SearchedLogStream

Creates a value of SearchedLogStream with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

slsLogStreamName - The name of the log stream.

slsSearchedCompletely - Indicates whether all the events in this log stream were searched.

slsLogStreamName :: Lens' SearchedLogStream (Maybe Text)

The name of the log stream.

slsSearchedCompletely :: Lens' SearchedLogStream (Maybe Bool)

Indicates whether all the events in this log stream were searched.

SubscriptionFilter

data SubscriptionFilter

Represents a subscription filter.

See: subscriptionFilter smart constructor.

subscriptionFilter :: SubscriptionFilter

Creates a value of SubscriptionFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sfCreationTime - The creation time of the subscription filter, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

sfFilterName - The name of the subscription filter.

sfDistribution - The method used to distribute log data to the destination, when the destination is an Amazon Kinesis stream.

sfDestinationARN - The Amazon Resource Name (ARN) of the destination.

sfLogGroupName - The name of the log group.

sfFilterPattern - Undocumented member.

sfRoleARN -

sfCreationTime :: Lens' SubscriptionFilter (Maybe Natural)

The creation time of the subscription filter, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

sfFilterName :: Lens' SubscriptionFilter (Maybe Text)

The name of the subscription filter.

sfDistribution :: Lens' SubscriptionFilter (Maybe Distribution)

The method used to distribute log data to the destination, when the destination is an Amazon Kinesis stream.

sfDestinationARN :: Lens' SubscriptionFilter (Maybe Text)

The Amazon Resource Name (ARN) of the destination.

sfLogGroupName :: Lens' SubscriptionFilter (Maybe Text)

The name of the log group.

sfFilterPattern :: Lens' SubscriptionFilter (Maybe Text)

Undocumented member.

sfRoleARN :: Lens' SubscriptionFilter (Maybe Text)