Network.AWS.CloudWatchLogs.CreateExportTask

Description

Creates an export task, which allows you to efficiently export data from a log group to an Amazon S3 bucket.

This is an asynchronous call. If all the required information is provided, this operation initiates an export task and responds with the ID of the task. After the task has started, you can use DescribeExportTasks to get the status of the export task. Each account can only have one active (RUNNING or PENDING ) export task at a time. To cancel an export task, use CancelExportTask .

You can export logs from multiple log groups or multiple time ranges to the same S3 bucket. To separate out log data for each export task, you can specify a prefix that will be used as the Amazon S3 key prefix for all exported objects.

Creating a Request

createExportTask

Arguments

:: Text
cetLogGroupName

-> Natural
cetFrom

-> Natural
cetTo

-> Text
cetDestination

-> CreateExportTask

Creates a value of CreateExportTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cetDestinationPrefix - The prefix used as the start of the key for every object exported. If you don't specify a value, the default is exportedlogs .

cetTaskName - The name of the export task.

cetLogStreamNamePrefix - Export only log streams that match the provided prefix. If you don't specify a value, no prefix filter is applied.

cetLogGroupName - The name of the log group.

cetFrom - The start time of the range for the request, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp earlier than this time are not exported.

cetTo - The end time of the range for the request, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

cetDestination - The name of S3 bucket for the exported log data. The bucket must be in the same AWS region.

data CreateExportTask

See: createExportTask smart constructor.

Request Lenses

cetDestinationPrefix :: Lens' CreateExportTask (Maybe Text)

The prefix used as the start of the key for every object exported. If you don't specify a value, the default is exportedlogs .

cetTaskName :: Lens' CreateExportTask (Maybe Text)

The name of the export task.

cetLogStreamNamePrefix :: Lens' CreateExportTask (Maybe Text)

Export only log streams that match the provided prefix. If you don't specify a value, no prefix filter is applied.

cetLogGroupName :: Lens' CreateExportTask Text

The name of the log group.

cetFrom :: Lens' CreateExportTask Natural

The start time of the range for the request, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp earlier than this time are not exported.

cetTo :: Lens' CreateExportTask Natural

The end time of the range for the request, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

cetDestination :: Lens' CreateExportTask Text

The name of S3 bucket for the exported log data. The bucket must be in the same AWS region.

Destructuring the Response

createExportTaskResponse

Arguments

:: Int
cetrsResponseStatus

-> CreateExportTaskResponse

Creates a value of CreateExportTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cetrsTaskId - The ID of the export task.

cetrsResponseStatus - -- | The response status code.

data CreateExportTaskResponse

See: createExportTaskResponse smart constructor.

Response Lenses

cetrsTaskId :: Lens' CreateExportTaskResponse (Maybe Text)

The ID of the export task.

cetrsResponseStatus :: Lens' CreateExportTaskResponse Int

- | The response status code.