Network.AWS.CloudWatchLogs.CreateLogGroup

Description

Creates a log group with the specified name.

You can create up to 5000 log groups per account.

You must use the following guidelines when naming a log group:

Log group names must be unique within a region for an AWS account.

Log group names can be between 1 and 512 characters long.

Log group names consist of the following characters: a-z, A-Z, 0-9, '_' (underscore), - (hyphen), / (forward slash), and . (period).

Creating a Request

createLogGroup

Arguments

:: Text
clgLogGroupName

-> CreateLogGroup

Creates a value of CreateLogGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clgTags - The key-value pairs to use for the tags.

clgLogGroupName - The name of the log group.

data CreateLogGroup

See: createLogGroup smart constructor.

Request Lenses

clgTags :: Lens' CreateLogGroup (HashMap Text Text)

The key-value pairs to use for the tags.

clgLogGroupName :: Lens' CreateLogGroup Text

The name of the log group.

Destructuring the Response

createLogGroupResponse :: CreateLogGroupResponse

Creates a value of CreateLogGroupResponse with the minimum fields required to make a request.

data CreateLogGroupResponse

See: createLogGroupResponse smart constructor.