Network.AWS.CloudWatchLogs.CreateLogStream

Description

Creates a new log stream in the specified log group. The name of the log stream must be unique within the log group. There is no limit on the number of log streams that can exist in a log group. You must use the following guidelines when naming a log stream: Log stream names can be between 1 and 512 characters long. The : colon character is not allowed.

http://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_CreateLogStream.html

Request

data CreateLogStream

Request constructor

createLogStream

Arguments

:: Text
clsLogGroupName

-> Text
clsLogStreamName

-> CreateLogStream

CreateLogStream constructor.

The fields accessible through corresponding lenses are:

clsLogGroupName :: Text

clsLogStreamName :: Text

Request lenses

clsLogGroupName :: Lens' CreateLogStream Text

clsLogStreamName :: Lens' CreateLogStream Text

Response

data CreateLogStreamResponse

Response constructor

createLogStreamResponse :: CreateLogStreamResponse

CreateLogStreamResponse constructor.