Network.AWS.CloudWatchLogs.DeleteDestination

Description

Deletes the specified destination, and eventually disables all the subscription filters that publish to it. This operation does not delete the physical resource encapsulated by the destination.

Creating a Request

deleteDestination

Arguments

:: Text
ddDestinationName

-> DeleteDestination

Creates a value of DeleteDestination with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddDestinationName - The name of the destination.

data DeleteDestination

See: deleteDestination smart constructor.

Request Lenses

ddDestinationName :: Lens' DeleteDestination Text

The name of the destination.

Destructuring the Response

deleteDestinationResponse :: DeleteDestinationResponse

Creates a value of DeleteDestinationResponse with the minimum fields required to make a request.

data DeleteDestinationResponse

See: deleteDestinationResponse smart constructor.