Network.AWS.CloudWatchLogs.DeleteLogGroup

Description

Deletes the log group with the specified name and permanently deletes all the archived log events associated with it.

http://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_DeleteLogGroup.html

Request

data DeleteLogGroup

Request constructor

deleteLogGroup

Arguments

:: Text
dlgLogGroupName

-> DeleteLogGroup

DeleteLogGroup constructor.

The fields accessible through corresponding lenses are:

dlgLogGroupName :: Text

Request lenses

dlgLogGroupName :: Lens' DeleteLogGroup Text

Response

data DeleteLogGroupResponse

Response constructor

deleteLogGroupResponse :: DeleteLogGroupResponse

DeleteLogGroupResponse constructor.