Network.AWS.CloudWatchLogs.DeleteSubscriptionFilter

Description

Deletes the specified subscription filter.

Creating a Request

deleteSubscriptionFilter

Arguments

:: Text
dLogGroupName

-> Text
dFilterName

-> DeleteSubscriptionFilter

Creates a value of DeleteSubscriptionFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dLogGroupName - The name of the log group.

dFilterName - The name of the subscription filter.

data DeleteSubscriptionFilter

See: deleteSubscriptionFilter smart constructor.

Request Lenses

dLogGroupName :: Lens' DeleteSubscriptionFilter Text

The name of the log group.

dFilterName :: Lens' DeleteSubscriptionFilter Text

The name of the subscription filter.

Destructuring the Response

deleteSubscriptionFilterResponse :: DeleteSubscriptionFilterResponse

Creates a value of DeleteSubscriptionFilterResponse with the minimum fields required to make a request.

data DeleteSubscriptionFilterResponse

See: deleteSubscriptionFilterResponse smart constructor.