Network.AWS.CloudWatchLogs.DescribeLogGroups

Description

Returns all the log groups that are associated with the AWS account making the request. The list returned in the response is ASCII-sorted by log group name.

By default, this operation returns up to 50 log groups. If there are more log groups to list, the response would contain a nextToken value in the response body. You can also limit the number of log groups returned in the response by specifying the limit parameter in the request.

http://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_DescribeLogGroups.html

Request

data DescribeLogGroups

Request constructor

describeLogGroups :: DescribeLogGroups

DescribeLogGroups constructor.

The fields accessible through corresponding lenses are:

dlgLimit :: Maybe Natural

dlgLogGroupNamePrefix :: Maybe Text

dlgNextToken :: Maybe Text

Request lenses

dlgLimit :: Lens' DescribeLogGroups (Maybe Natural)

The maximum number of items returned in the response. If you don't specify a value, the request would return up to 50 items.

dlgLogGroupNamePrefix :: Lens' DescribeLogGroups (Maybe Text)

dlgNextToken :: Lens' DescribeLogGroups (Maybe Text)

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous DescribeLogGroups request.

Response

data DescribeLogGroupsResponse

Response constructor

describeLogGroupsResponse :: DescribeLogGroupsResponse

DescribeLogGroupsResponse constructor.

The fields accessible through corresponding lenses are:

dlgrLogGroups :: [LogGroup]

dlgrNextToken :: Maybe Text

Response lenses

dlgrLogGroups :: Lens' DescribeLogGroupsResponse [LogGroup]

dlgrNextToken :: Lens' DescribeLogGroupsResponse (Maybe Text)