Network.AWS.CloudWatchLogs.DescribeLogStreams

Description

Returns all the log streams that are associated with the specified log group. The list returned in the response is ASCII-sorted by log stream name.

By default, this operation returns up to 50 log streams. If there are more log streams to list, the response would contain a nextToken value in the response body. You can also limit the number of log streams returned in the response by specifying the limit parameter in the request. This operation has a limit of five transactions per second, after which transactions are throttled.

http://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_DescribeLogStreams.html

Request

data DescribeLogStreams

Request constructor

describeLogStreams

Arguments

:: Text
dls1LogGroupName

-> DescribeLogStreams

DescribeLogStreams constructor.

The fields accessible through corresponding lenses are:

dls1Descending :: Maybe Bool

dls1Limit :: Maybe Natural

dls1LogGroupName :: Text

dls1LogStreamNamePrefix :: Maybe Text

dls1NextToken :: Maybe Text

dls1OrderBy :: Maybe OrderBy

Request lenses

dls1Descending :: Lens' DescribeLogStreams (Maybe Bool)

If set to true, results are returned in descending order. If you don't specify a value or set it to false, results are returned in ascending order.

dls1Limit :: Lens' DescribeLogStreams (Maybe Natural)

The maximum number of items returned in the response. If you don't specify a value, the request would return up to 50 items.

dls1LogGroupName :: Lens' DescribeLogStreams Text

dls1LogStreamNamePrefix :: Lens' DescribeLogStreams (Maybe Text)

Will only return log streams that match the provided logStreamNamePrefix. If you don't specify a value, no prefix filter is applied.

dls1NextToken :: Lens' DescribeLogStreams (Maybe Text)

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous DescribeLogStreams request.

dls1OrderBy :: Lens' DescribeLogStreams (Maybe OrderBy)

Specifies what to order the returned log streams by. Valid arguments are LogStreamName or LastEventTime. If you don't specify a value, results are ordered by LogStreamName. If LastEventTime is chosen, the request cannot also contain a logStreamNamePrefix.

Response

data DescribeLogStreamsResponse

Response constructor

describeLogStreamsResponse :: DescribeLogStreamsResponse

DescribeLogStreamsResponse constructor.

The fields accessible through corresponding lenses are:

dlsrLogStreams :: [LogStream]

dlsrNextToken :: Maybe Text

Response lenses

dlsrLogStreams :: Lens' DescribeLogStreamsResponse [LogStream]

dlsrNextToken :: Lens' DescribeLogStreamsResponse (Maybe Text)