Network.AWS.CloudWatchLogs.DescribeMetricFilters

Description

Returns all the metrics filters associated with the specified log group. The list returned in the response is ASCII-sorted by filter name.

By default, this operation returns up to 50 metric filters. If there are more metric filters to list, the response would contain a nextToken value in the response body. You can also limit the number of metric filters returned in the response by specifying the limit parameter in the request.

http://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_DescribeMetricFilters.html

Request

data DescribeMetricFilters

Request constructor

describeMetricFilters

Arguments

:: Text
dmfLogGroupName

-> DescribeMetricFilters

DescribeMetricFilters constructor.

The fields accessible through corresponding lenses are:

dmfFilterNamePrefix :: Maybe Text

dmfLimit :: Maybe Natural

dmfLogGroupName :: Text

dmfNextToken :: Maybe Text

Request lenses

dmfFilterNamePrefix :: Lens' DescribeMetricFilters (Maybe Text)

dmfLimit :: Lens' DescribeMetricFilters (Maybe Natural)

The maximum number of items returned in the response. If you don't specify a value, the request would return up to 50 items.

dmfLogGroupName :: Lens' DescribeMetricFilters Text

dmfNextToken :: Lens' DescribeMetricFilters (Maybe Text)

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous DescribeMetricFilters request.

Response

data DescribeMetricFiltersResponse

Response constructor

describeMetricFiltersResponse :: DescribeMetricFiltersResponse

DescribeMetricFiltersResponse constructor.

The fields accessible through corresponding lenses are:

dmfrMetricFilters :: [MetricFilter]

dmfrNextToken :: Maybe Text

Response lenses

dmfrMetricFilters :: Lens' DescribeMetricFiltersResponse [MetricFilter]

dmfrNextToken :: Lens' DescribeMetricFiltersResponse (Maybe Text)