Network.AWS.CloudWatchLogs.FilterLogEvents

Description

Lists log events from the specified log group. You can list all the log events or filter the results using a filter pattern, a time range, and the name of the log stream.

By default, this operation returns as many log events as can fit in 1MB (up to 10,000 log events), or all the events found within the time range that you specify. If the results include a token, then there are more log events available, and you can get additional results by specifying the token in a subsequent call.

This operation returns paginated results.

Creating a Request

filterLogEvents

Arguments

:: Text
fleLogGroupName

-> FilterLogEvents

Creates a value of FilterLogEvents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fleStartTime - The start of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp prior to this time are not returned.

fleNextToken - The token for the next set of events to return. (You received this token from a previous call.)

fleLogStreamNames - Optional list of log stream names.

fleEndTime - The end of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not returned.

fleLimit - The maximum number of events to return. The default is 10,000 events.

fleFilterPattern - The filter pattern to use. If not provided, all the events are matched.

fleInterleaved - If the value is true, the operation makes a best effort to provide responses that contain events from multiple log streams within the log group interleaved in a single response. If the value is false all the matched log events in the first log stream are searched first, then those in the next log stream, and so on. The default is false.

fleLogGroupName - The name of the log group.

data FilterLogEvents

See: filterLogEvents smart constructor.

Request Lenses

fleStartTime :: Lens' FilterLogEvents (Maybe Natural)

The start of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp prior to this time are not returned.

fleNextToken :: Lens' FilterLogEvents (Maybe Text)

The token for the next set of events to return. (You received this token from a previous call.)

fleLogStreamNames :: Lens' FilterLogEvents (Maybe (NonEmpty Text))

Optional list of log stream names.

fleEndTime :: Lens' FilterLogEvents (Maybe Natural)

The end of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not returned.

fleLimit :: Lens' FilterLogEvents (Maybe Natural)

The maximum number of events to return. The default is 10,000 events.

fleFilterPattern :: Lens' FilterLogEvents (Maybe Text)

The filter pattern to use. If not provided, all the events are matched.

fleInterleaved :: Lens' FilterLogEvents (Maybe Bool)

If the value is true, the operation makes a best effort to provide responses that contain events from multiple log streams within the log group interleaved in a single response. If the value is false all the matched log events in the first log stream are searched first, then those in the next log stream, and so on. The default is false.

fleLogGroupName :: Lens' FilterLogEvents Text

The name of the log group.

Destructuring the Response

filterLogEventsResponse

Arguments

:: Int
flersResponseStatus

-> FilterLogEventsResponse

Creates a value of FilterLogEventsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

flersSearchedLogStreams - Indicates which log streams have been searched and whether each has been searched completely.

flersNextToken - The token to use when requesting the next set of items. The token expires after 24 hours.

flersEvents - The matched events.

flersResponseStatus - -- | The response status code.

data FilterLogEventsResponse

See: filterLogEventsResponse smart constructor.

Response Lenses

flersSearchedLogStreams :: Lens' FilterLogEventsResponse [SearchedLogStream]

Indicates which log streams have been searched and whether each has been searched completely.

flersNextToken :: Lens' FilterLogEventsResponse (Maybe Text)

The token to use when requesting the next set of items. The token expires after 24 hours.

flersEvents :: Lens' FilterLogEventsResponse [FilteredLogEvent]

The matched events.

flersResponseStatus :: Lens' FilterLogEventsResponse Int

- | The response status code.