Network.AWS.CloudWatchLogs.PutDestination

Description

Creates or updates a destination. A destination encapsulates a physical resource (such as a Kinesis stream) and enables you to subscribe to a real-time stream of log events of a different account, ingested using PutLogEvents . Currently, the only supported physical resource is a Amazon Kinesis stream belonging to the same account as the destination.

A destination controls what is written to its Amazon Kinesis stream through an access policy. By default, PutDestination does not set any access policy with the destination, which means a cross-account user cannot call PutSubscriptionFilter against this destination. To enable this, the destination owner must call PutDestinationPolicy after PutDestination .

Creating a Request

putDestination

Arguments

:: Text
pdDestinationName

-> Text
pdTargetARN

-> Text
pdRoleARN

-> PutDestination

Creates a value of PutDestination with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pdDestinationName - A name for the destination.

pdTargetARN - The ARN of an Amazon Kinesis stream to deliver matching log events to.

pdRoleARN - The ARN of an IAM role that grants CloudWatch Logs permissions to call Amazon Kinesis PutRecord on the destination stream.

data PutDestination

See: putDestination smart constructor.

Request Lenses

pdDestinationName :: Lens' PutDestination Text

A name for the destination.

pdTargetARN :: Lens' PutDestination Text

The ARN of an Amazon Kinesis stream to deliver matching log events to.

pdRoleARN :: Lens' PutDestination Text

The ARN of an IAM role that grants CloudWatch Logs permissions to call Amazon Kinesis PutRecord on the destination stream.

Destructuring the Response

putDestinationResponse

Arguments

:: Int
pdrsResponseStatus

-> PutDestinationResponse

Creates a value of PutDestinationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pdrsDestination - The destination.

pdrsResponseStatus - -- | The response status code.

data PutDestinationResponse

See: putDestinationResponse smart constructor.

Response Lenses

pdrsDestination :: Lens' PutDestinationResponse (Maybe Destination)

The destination.

pdrsResponseStatus :: Lens' PutDestinationResponse Int

- | The response status code.