Network.AWS.CloudWatchLogs.PutLogEvents

Description

Uploads a batch of log events to the specified log stream. Every PutLogEvents request must include the sequenceToken obtained from the response of the previous request. An upload in a newly created log stream does not require a sequenceToken. The batch of events must satisfy the following constraints: The maximum batch size is 32,768 bytes, and this size is calculated as the sum of all event messages in UTF-8, plus 26 bytes for each log event. None of the log events in the batch can be more than 2 hours in the future. None of the log events in the batch can be older than 14 days or the retention period of the log group. The log events in the batch must be in chronological ordered by their timestamp. The maximum number of log events in a batch is 1,000.

http://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_PutLogEvents.html

Request

data PutLogEvents

Request constructor

putLogEvents

Arguments

:: Text
pleLogGroupName

-> Text
pleLogStreamName

-> NonEmpty InputLogEvent
pleLogEvents

-> PutLogEvents

PutLogEvents constructor.

The fields accessible through corresponding lenses are:

pleLogEvents :: NonEmpty InputLogEvent

pleLogGroupName :: Text

pleLogStreamName :: Text

pleSequenceToken :: Maybe Text

Request lenses

pleLogEvents :: Lens' PutLogEvents (NonEmpty InputLogEvent)

pleLogGroupName :: Lens' PutLogEvents Text

pleLogStreamName :: Lens' PutLogEvents Text

pleSequenceToken :: Lens' PutLogEvents (Maybe Text)

A string token that must be obtained from the response of the previous PutLogEvents request.

Response

data PutLogEventsResponse

Response constructor

putLogEventsResponse :: PutLogEventsResponse

PutLogEventsResponse constructor.

The fields accessible through corresponding lenses are:

plerNextSequenceToken :: Maybe Text

Response lenses

plerNextSequenceToken :: Lens' PutLogEventsResponse (Maybe Text)