Network.AWS.CloudWatchLogs.PutMetricFilter

Description

Creates or updates a metric filter and associates it with the specified log group. Metric filters allow you to configure rules to extract metric data from log events ingested through PutLogEvents requests.

http://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_PutMetricFilter.html

Request

data PutMetricFilter

Request constructor

putMetricFilter

Arguments

:: Text
pmfLogGroupName

-> Text
pmfFilterName

-> Text
pmfFilterPattern

-> NonEmpty MetricTransformation
pmfMetricTransformations

-> PutMetricFilter

PutMetricFilter constructor.

The fields accessible through corresponding lenses are:

pmfFilterName :: Text

pmfFilterPattern :: Text

pmfLogGroupName :: Text

pmfMetricTransformations :: NonEmpty MetricTransformation

Request lenses

pmfFilterName :: Lens' PutMetricFilter Text

pmfFilterPattern :: Lens' PutMetricFilter Text

pmfLogGroupName :: Lens' PutMetricFilter Text

pmfMetricTransformations :: Lens' PutMetricFilter (NonEmpty MetricTransformation)

Response

data PutMetricFilterResponse

Response constructor

putMetricFilterResponse :: PutMetricFilterResponse

PutMetricFilterResponse constructor.