Network.AWS.CloudWatchLogs.PutRetentionPolicy

Description

Sets the retention of the specified log group. A retention policy allows you to configure the number of days you want to retain log events in the specified log group. Creates or updates a 30 day retention policy for a log group The following is an example of a PutRetentionPolicy request and response. POST HTTP1.1 Host: logs.. X-Amz-Date: Authorization: AWS4-HMAC-SHA256 Credential=, SignedHeaders=content-type;date;host;user-agent;x-amz-date;x-amz-target;x-amzn-requestid, Signature= User-Agent: Accept: application/json Content-Type: application/x-amz-json-1.1 Content-Length: Connection: Keep-Alive]]> X-Amz-Target: Logs_20140328.PutRetentionPolicy { "logGroupName": "exampleLogGroupName", "retentionInDays": 30 } HTTP/1.1 200 OK x-amzn-RequestId: Content-Type: application/x-amz-json-1.1 Content-Length: Date: ]]>.

Request

data PutRetentionPolicy

Request constructor

putRetentionPolicy

Arguments

:: Text
prpLogGroupName

-> Integer
prpRetentionInDays

-> PutRetentionPolicy

Smart constructor for the minimum required parameters to construct a valid PutRetentionPolicy request.

The fields accessible through corresponding lenses are:

LogGroupName :: Text

RetentionInDays :: Integer

Request lenses

prpLogGroupName :: Lens' PutRetentionPolicy Text

prpRetentionInDays :: Lens' PutRetentionPolicy Integer

Specifies the number of days you want to retain log events in the specified log group. Possible values are: 1, 3, 5, 7, 14, 30, 60, 90, 120, 150, 180, 365, 400, 547, 730.

Response

data PutRetentionPolicyResponse

Response constructor

putRetentionPolicyResponse :: PutRetentionPolicyResponse

Smart constructor for the minimum required parameters to construct a valid PutRetentionPolicyResponse response.

This constructor is provided for convenience and testing purposes.