Network.AWS.CloudWatchLogs.PutSubscriptionFilter

Description

Creates or updates a subscription filter and associates it with the specified log group. Subscription filters allow you to subscribe to a real-time stream of log events ingested through PutLogEvents and have them delivered to a specific destination. Currently, the supported destinations are:

An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery.

A logical destination that belongs to a different account, for cross-account delivery.

An Amazon Kinesis Firehose stream that belongs to the same account as the subscription filter, for same-account delivery.

An AWS Lambda function that belongs to the same account as the subscription filter, for same-account delivery.

There can only be one subscription filter associated with a log group. If you are updating an existing filter, you must specify the correct name in filterName . Otherwise, the call will fail because you cannot associate a second filter with a log group.

Creating a Request

putSubscriptionFilter

Arguments

:: Text
psfLogGroupName

-> Text
psfFilterName

-> Text
psfFilterPattern

-> Text
psfDestinationARN

-> PutSubscriptionFilter

Creates a value of PutSubscriptionFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

psfDistribution - The method used to distribute log data to the destination, when the destination is an Amazon Kinesis stream. By default, log data is grouped by log stream. For a more even distribution, you can group log data randomly.

psfRoleARN - The ARN of an IAM role that grants CloudWatch Logs permissions to deliver ingested log events to the destination stream. You don't need to provide the ARN when you are working with a logical destination for cross-account delivery.

psfLogGroupName - The name of the log group.

psfFilterName - A name for the subscription filter. If you are updating an existing filter, you must specify the correct name in filterName . Otherwise, the call will fail because you cannot associate a second filter with a log group. To find the name of the filter currently associated with a log group, use DescribeSubscriptionFilters .

psfFilterPattern - A filter pattern for subscribing to a filtered stream of log events.

psfDestinationARN - The ARN of the destination to deliver matching log events to. Currently, the supported destinations are: * An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery. * A logical destination (specified using an ARN) belonging to a different account, for cross-account delivery. * An Amazon Kinesis Firehose stream belonging to the same account as the subscription filter, for same-account delivery. * An AWS Lambda function belonging to the same account as the subscription filter, for same-account delivery.

data PutSubscriptionFilter

See: putSubscriptionFilter smart constructor.

Request Lenses

psfDistribution :: Lens' PutSubscriptionFilter (Maybe Distribution)

The method used to distribute log data to the destination, when the destination is an Amazon Kinesis stream. By default, log data is grouped by log stream. For a more even distribution, you can group log data randomly.

psfRoleARN :: Lens' PutSubscriptionFilter (Maybe Text)

The ARN of an IAM role that grants CloudWatch Logs permissions to deliver ingested log events to the destination stream. You don't need to provide the ARN when you are working with a logical destination for cross-account delivery.

psfLogGroupName :: Lens' PutSubscriptionFilter Text

The name of the log group.

psfFilterName :: Lens' PutSubscriptionFilter Text

A name for the subscription filter. If you are updating an existing filter, you must specify the correct name in filterName . Otherwise, the call will fail because you cannot associate a second filter with a log group. To find the name of the filter currently associated with a log group, use DescribeSubscriptionFilters .

psfFilterPattern :: Lens' PutSubscriptionFilter Text

A filter pattern for subscribing to a filtered stream of log events.

psfDestinationARN :: Lens' PutSubscriptionFilter Text

The ARN of the destination to deliver matching log events to. Currently, the supported destinations are: * An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery. * A logical destination (specified using an ARN) belonging to a different account, for cross-account delivery. * An Amazon Kinesis Firehose stream belonging to the same account as the subscription filter, for same-account delivery. * An AWS Lambda function belonging to the same account as the subscription filter, for same-account delivery.

Destructuring the Response

putSubscriptionFilterResponse :: PutSubscriptionFilterResponse

Creates a value of PutSubscriptionFilterResponse with the minimum fields required to make a request.

data PutSubscriptionFilterResponse

See: putSubscriptionFilterResponse smart constructor.