Network.AWS.CloudWatchLogs.TagLogGroup

Description

Adds or updates the specified tags for the specified log group.

To list the tags for a log group, use ListTagsLogGroup . To remove tags, use UntagLogGroup .

For more information about tags, see Tag Log Groups in Amazon CloudWatch Logs in the Amazon CloudWatch Logs User Guide .

Creating a Request

tagLogGroup

Arguments

:: Text
tlgLogGroupName

-> TagLogGroup

Creates a value of TagLogGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tlgLogGroupName - The name of the log group.

tlgTags - The key-value pairs to use for the tags.

data TagLogGroup

See: tagLogGroup smart constructor.

Request Lenses

tlgLogGroupName :: Lens' TagLogGroup Text

The name of the log group.

tlgTags :: Lens' TagLogGroup (HashMap Text Text)

The key-value pairs to use for the tags.

Destructuring the Response

tagLogGroupResponse :: TagLogGroupResponse

Creates a value of TagLogGroupResponse with the minimum fields required to make a request.

data TagLogGroupResponse

See: tagLogGroupResponse smart constructor.