Network.AWS.CloudWatchLogs.Types

Description

Amazon CloudWatch Logs enables you to monitor, store, and access your system, application, and custom log files.

Service

data CloudWatchLogs

Supported version (2014-03-28) of the Amazon CloudWatch Logs service.

Errors

data CloudWatchLogsError

A sum type representing possible errors returned by the CloudWatchLogs service.

These typically include HTTPExceptions thrown by the underlying HTTP mechanisms, serialisation errors, and typed errors as specified by the service description where applicable.

Constructors

CloudWatchLogsClient HttpException

CloudWatchLogsSerializer String

CloudWatchLogsService String

DataAlreadyAcceptedException

Fields

_daaeExpectedSequenceToken :: Maybe Text

InvalidParameterException
Returned if a parameter of the request is incorrectly specified.

InvalidSequenceTokenException

Fields

_isteExpectedSequenceToken :: Maybe Text

LimitExceededException
Returned if you have reached the maximum number of resources that can be created.

OperationAbortedException
Returned if multiple requests to update the same resource were in conflict.

ResourceAlreadyExistsException
Returned if the specified resource already exists.

ResourceNotFoundException
Returned if the specified resource does not exist.

ServiceUnavailableException
Returned if the service cannot complete the request.

_CloudWatchLogsClient :: Prism' CloudWatchLogsError HttpException

See: CloudWatchLogsClient

_CloudWatchLogsSerializer :: Prism' CloudWatchLogsError String

See: CloudWatchLogsSerializer

_CloudWatchLogsService :: Prism' CloudWatchLogsError String

See: CloudWatchLogsService

_DataAlreadyAcceptedException :: Prism' CloudWatchLogsError (Maybe Text)

See: DataAlreadyAcceptedException

_InvalidParameterException :: Prism' CloudWatchLogsError ()

Returned if a parameter of the request is incorrectly specified.

See: InvalidParameterException

_InvalidSequenceTokenException :: Prism' CloudWatchLogsError (Maybe Text)

See: InvalidSequenceTokenException

_LimitExceededException :: Prism' CloudWatchLogsError ()

Returned if you have reached the maximum number of resources that can be created.

See: LimitExceededException

_OperationAbortedException :: Prism' CloudWatchLogsError ()

Returned if multiple requests to update the same resource were in conflict.

See: OperationAbortedException

_ResourceAlreadyExistsException :: Prism' CloudWatchLogsError ()

Returned if the specified resource already exists.

See: ResourceAlreadyExistsException

_ResourceNotFoundException :: Prism' CloudWatchLogsError ()

Returned if the specified resource does not exist.

See: ResourceNotFoundException

_ServiceUnavailableException :: Prism' CloudWatchLogsError ()

Returned if the service cannot complete the request.

See: ServiceUnavailableException

InputLogEvent

data InputLogEvent

A log event is a record of some activity that was recorded by the application or resource being monitored. The log event record that Amazon CloudWatch Logs understands contains two properties: the timestamp of when the event occurred, and the raw event message.

inputLogEvent

Arguments

:: Integer
ileTimestamp

-> Text
ileMessage

-> InputLogEvent

Smart constructor for the minimum required fields to construct a valid InputLogEvent data type to populate a request.

The fields accessible through corresponding lenses are:

Timestamp :: Integer

Message :: Text

ileTimestamp :: Lens' InputLogEvent Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

ileMessage :: Lens' InputLogEvent Text

LogGroup

data LogGroup

logGroup :: LogGroup

Smart constructor for the minimum required fields to construct a valid LogGroup data type.

LogGroup is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

LogGroupName :: Maybe Text

CreationTime :: Maybe Integer

RetentionInDays :: Maybe Integer

MetricFilterCount :: Maybe Integer

Arn :: Maybe Text

StoredBytes :: Maybe Integer

lgLogGroupName :: Lens' LogGroup (Maybe Text)

lgCreationTime :: Lens' LogGroup (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

lgRetentionInDays :: Lens' LogGroup (Maybe Integer)

Specifies the number of days you want to retain log events in the specified log group. Possible values are: 1, 3, 5, 7, 14, 30, 60, 90, 120, 150, 180, 365, 400, 547, 730.

lgMetricFilterCount :: Lens' LogGroup (Maybe Integer)

The number of metric filters associated with the log group.

lgArn :: Lens' LogGroup (Maybe Text)

lgStoredBytes :: Lens' LogGroup (Maybe Integer)

LogStream

data LogStream

A log stream is sequence of log events that share the same emitter.

logStream :: LogStream

Smart constructor for the minimum required fields to construct a valid LogStream data type.

LogStream is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

LogStreamName :: Maybe Text

CreationTime :: Maybe Integer

FirstEventTimestamp :: Maybe Integer

LastEventTimestamp :: Maybe Integer

LastIngestionTime :: Maybe Integer

UploadSequenceToken :: Maybe Text

Arn :: Maybe Text

StoredBytes :: Maybe Integer

lsLogStreamName :: Lens' LogStream (Maybe Text)

lsCreationTime :: Lens' LogStream (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

lsFirstEventTimestamp :: Lens' LogStream (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

lsLastEventTimestamp :: Lens' LogStream (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

lsLastIngestionTime :: Lens' LogStream (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

lsUploadSequenceToken :: Lens' LogStream (Maybe Text)

A string token used for making PutLogEvents requests. A sequenceToken can only be used once, and PutLogEvents requests must include the sequenceToken obtained from the response of the previous request.

lsArn :: Lens' LogStream (Maybe Text)

lsStoredBytes :: Lens' LogStream (Maybe Integer)

MetricFilter

data MetricFilter

Metric filters can be used to express how Amazon CloudWatch Logs would extract metric observations from ingested log events and transform them to metric data in a CloudWatch metric.

metricFilter :: MetricFilter

Smart constructor for the minimum required fields to construct a valid MetricFilter data type.

MetricFilter is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

FilterName :: Maybe Text

FilterPattern :: Maybe Text

MetricTransformations :: Maybe (List1 MetricTransformation)

CreationTime :: Maybe Integer

mfFilterName :: Lens' MetricFilter (Maybe Text)

The name of the metric filter.

mfFilterPattern :: Lens' MetricFilter (Maybe Text)

mfMetricTransformations :: Lens' MetricFilter (Maybe (List1 MetricTransformation))

mfCreationTime :: Lens' MetricFilter (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

MetricFilterMatchRecord

data MetricFilterMatchRecord

metricFilterMatchRecord :: MetricFilterMatchRecord

Smart constructor for the minimum required fields to construct a valid MetricFilterMatchRecord data type.

MetricFilterMatchRecord is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

EventNumber :: Maybe Integer

EventMessage :: Maybe Text

ExtractedValues :: Map Text Text

mfmrEventNumber :: Lens' MetricFilterMatchRecord (Maybe Integer)

mfmrEventMessage :: Lens' MetricFilterMatchRecord (Maybe Text)

mfmrExtractedValues :: Lens' MetricFilterMatchRecord (Map Text Text)

MetricTransformation

data MetricTransformation

metricTransformation

Arguments

:: Text
mtMetricName

-> Text
mtMetricNamespace

-> Text
mtMetricValue

-> MetricTransformation

Smart constructor for the minimum required fields to construct a valid MetricTransformation data type to populate a request.

The fields accessible through corresponding lenses are:

MetricName :: Text

MetricNamespace :: Text

MetricValue :: Text

mtMetricName :: Lens' MetricTransformation Text

mtMetricNamespace :: Lens' MetricTransformation Text

mtMetricValue :: Lens' MetricTransformation Text

OutputLogEvent

data OutputLogEvent

outputLogEvent :: OutputLogEvent

Smart constructor for the minimum required fields to construct a valid OutputLogEvent data type.

OutputLogEvent is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Timestamp :: Maybe Integer

Message :: Maybe Text

IngestionTime :: Maybe Integer

oleTimestamp :: Lens' OutputLogEvent (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

oleMessage :: Lens' OutputLogEvent (Maybe Text)

oleIngestionTime :: Lens' OutputLogEvent (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.