Network.AWS.CloudWatchLogs.UntagLogGroup

Description

Removes the specified tags from the specified log group.

To list the tags for a log group, use ListTagsLogGroup . To add tags, use UntagLogGroup .

Creating a Request

untagLogGroup

Arguments

:: Text
ulgLogGroupName

-> NonEmpty Text
ulgTags

-> UntagLogGroup

Creates a value of UntagLogGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ulgLogGroupName - The name of the log group.

ulgTags - The tag keys. The corresponding tags are removed from the log group.

data UntagLogGroup

See: untagLogGroup smart constructor.

Request Lenses

ulgLogGroupName :: Lens' UntagLogGroup Text

The name of the log group.

ulgTags :: Lens' UntagLogGroup (NonEmpty Text)

The tag keys. The corresponding tags are removed from the log group.

Destructuring the Response

untagLogGroupResponse :: UntagLogGroupResponse

Creates a value of UntagLogGroupResponse with the minimum fields required to make a request.

data UntagLogGroupResponse

See: untagLogGroupResponse smart constructor.