Network.AWS.CloudWatchLogs.V2014_03_28

Description

Amazon CloudWatch Logs enables you to monitor, store, and access your system, application, and custom log files.

Documentation

data CreateLogGroup

Constructors

CreateLogGroup

Fields

_clgrLogGroupName :: Text

clgrLogGroupName :: Iso' CreateLogGroup Text

data CreateLogGroupResponse

Constructors

CreateLogGroupResponse

data CreateLogStream

Constructors

CreateLogStream

Fields

_clsrLogGroupName :: Text

_clsrLogStreamName :: Text

clsrLogStreamName :: Lens' CreateLogStream Text

clsrLogGroupName :: Lens' CreateLogStream Text

data CreateLogStreamResponse

Constructors

CreateLogStreamResponse

data DeleteLogGroup

Constructors

DeleteLogGroup

Fields

_dlgtLogGroupName :: Text

dlgtLogGroupName :: Iso' DeleteLogGroup Text

data DeleteLogGroupResponse

Constructors

DeleteLogGroupResponse

data DeleteLogStream

Constructors

DeleteLogStream

Fields

_dlsrLogGroupName :: Text

_dlsrLogStreamName :: Text

dlsrLogStreamName :: Lens' DeleteLogStream Text

dlsrLogGroupName :: Lens' DeleteLogStream Text

data DeleteLogStreamResponse

Constructors

DeleteLogStreamResponse

data DeleteMetricFilter

Constructors

DeleteMetricFilter

Fields

_dmftFilterName :: Text

The name of the metric filter.

_dmftLogGroupName :: Text

dmftLogGroupName :: Lens' DeleteMetricFilter Text

dmftFilterName :: Lens' DeleteMetricFilter Text

data DeleteMetricFilterResponse

Constructors

DeleteMetricFilterResponse

data DeleteRetentionPolicy

Constructors

DeleteRetentionPolicy

Fields

_drprLogGroupName :: Text

drprLogGroupName :: Iso' DeleteRetentionPolicy Text

data DeleteRetentionPolicyResponse

Constructors

DeleteRetentionPolicyResponse

describeLogGroups :: DescribeLogGroups

Minimum specification for a DescribeLogGroups request.

data DescribeLogGroups

Constructors

DescribeLogGroups

Fields

_dlgrLimit :: Maybe Integer

The maximum number of items returned in the response. If you don't specify a value, the request would return up to 50 items.

_dlgrLogGroupNamePrefix :: Maybe Text

_dlgrNextToken :: Maybe Text

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous DescribeLogGroups request.

dlgrNextToken :: Lens' DescribeLogGroups (Maybe Text)

dlgrLogGroupNamePrefix :: Lens' DescribeLogGroups (Maybe Text)

dlgrLimit :: Lens' DescribeLogGroups (Maybe Integer)

data DescribeLogGroupsResponse

Constructors

DescribeLogGroupsResponse

Fields

_dlgsLogGroups :: [LogGroup]

A list of log groups.

_dlgsNextToken :: Maybe Text

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous request. The token expires after 24 hours.

dlgsNextToken :: Lens' DescribeLogGroupsResponse (Maybe Text)

dlgsLogGroups :: Lens' DescribeLogGroupsResponse [LogGroup]

describeLogStreams

Arguments

:: Text
_dlssLogGroupName

-> DescribeLogStreams

Minimum specification for a DescribeLogStreams request.

data DescribeLogStreams

Constructors

DescribeLogStreams

Fields

_dlssLogGroupName :: Text

_dlssLimit :: Maybe Integer

The maximum number of items returned in the response. If you don't specify a value, the request would return up to 50 items.

_dlssLogStreamNamePrefix :: Maybe Text

_dlssNextToken :: Maybe Text

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous DescribeLogStreams request.

dlssNextToken :: Lens' DescribeLogStreams (Maybe Text)

dlssLogStreamNamePrefix :: Lens' DescribeLogStreams (Maybe Text)

dlssLogGroupName :: Lens' DescribeLogStreams Text

dlssLimit :: Lens' DescribeLogStreams (Maybe Integer)

data DescribeLogStreamsResponse

Constructors

DescribeLogStreamsResponse

Fields

_dlstLogStreams :: [LogStream]

A list of log streams.

_dlstNextToken :: Maybe Text

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous request. The token expires after 24 hours.

dlstNextToken :: Lens' DescribeLogStreamsResponse (Maybe Text)

dlstLogStreams :: Lens' DescribeLogStreamsResponse [LogStream]

describeMetricFilters

Arguments

:: Text
_dmfrLogGroupName

-> DescribeMetricFilters

Minimum specification for a DescribeMetricFilters request.

data DescribeMetricFilters

Constructors

DescribeMetricFilters

Fields

_dmfrLogGroupName :: Text

_dmfrLimit :: Maybe Integer

The maximum number of items returned in the response. If you don't specify a value, the request would return up to 50 items.

_dmfrFilterNamePrefix :: Maybe Text

The name of the metric filter.

_dmfrNextToken :: Maybe Text

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous DescribeMetricFilters request.

dmfrNextToken :: Lens' DescribeMetricFilters (Maybe Text)

dmfrLogGroupName :: Lens' DescribeMetricFilters Text

dmfrLimit :: Lens' DescribeMetricFilters (Maybe Integer)

dmfrFilterNamePrefix :: Lens' DescribeMetricFilters (Maybe Text)

data DescribeMetricFiltersResponse

Constructors

DescribeMetricFiltersResponse

Fields

_dmfsMetricFilters :: [MetricFilter]

_dmfsNextToken :: Maybe Text

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous request. The token expires after 24 hours.

dmfsNextToken :: Lens' DescribeMetricFiltersResponse (Maybe Text)

dmfsMetricFilters :: Lens' DescribeMetricFiltersResponse [MetricFilter]

getLogEvents

Arguments

:: Text
_glerLogGroupName

-> Text
_glerLogStreamName

-> GetLogEvents

Minimum specification for a GetLogEvents request.

data GetLogEvents

Constructors

GetLogEvents

Fields

_glerLogGroupName :: Text

_glerLogStreamName :: Text

_glerLimit :: Maybe Integer

The maximum number of log events returned in the response. If you don't specify a value, the request would return as much log events as can fit in a response size of 1MB, up to 10,000 log events.

_glerNextToken :: Maybe Text

A string token used for pagination that points to the next page of results. It must be a value obtained from the nextForwardToken or nextBackwardToken fields in the response of the previous GetLogEvents request.

_glerStartFromHead :: Maybe Bool

_glerStartTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_glerEndTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

glerStartTime :: Lens' GetLogEvents (Maybe Integer)

glerStartFromHead :: Lens' GetLogEvents (Maybe Bool)

glerNextToken :: Lens' GetLogEvents (Maybe Text)

glerLogStreamName :: Lens' GetLogEvents Text

glerLogGroupName :: Lens' GetLogEvents Text

glerLimit :: Lens' GetLogEvents (Maybe Integer)

glerEndTime :: Lens' GetLogEvents (Maybe Integer)

data GetLogEventsResponse

Constructors

GetLogEventsResponse

Fields

_glesNextBackwardToken :: Maybe Text

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous request. The token expires after 24 hours.

_glesNextForwardToken :: Maybe Text

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous request. The token expires after 24 hours.

_glesEvents :: [OutputLogEvent]

glesNextForwardToken :: Lens' GetLogEventsResponse (Maybe Text)

glesNextBackwardToken :: Lens' GetLogEventsResponse (Maybe Text)

glesEvents :: Lens' GetLogEventsResponse [OutputLogEvent]

putLogEvents

Arguments

:: [InputLogEvent]
_plerLogEvents

-> Text
_plerLogGroupName

-> Text
_plerLogStreamName

-> PutLogEvents

Minimum specification for a PutLogEvents request.

data PutLogEvents

Constructors

PutLogEvents

Fields

_plerLogEvents :: [InputLogEvent]

A list of events belonging to a log stream.

_plerLogGroupName :: Text

_plerLogStreamName :: Text

_plerSequenceToken :: Maybe Text

A string token that must be obtained from the response of the previous PutLogEvents request.

plerSequenceToken :: Lens' PutLogEvents (Maybe Text)

plerLogStreamName :: Lens' PutLogEvents Text

plerLogGroupName :: Lens' PutLogEvents Text

plerLogEvents :: Lens' PutLogEvents [InputLogEvent]

data PutLogEventsResponse

Constructors

PutLogEventsResponse

Fields

_plesNextSequenceToken :: Maybe Text

A string token used for making PutLogEvents requests. A sequenceToken can only be used once, and PutLogEvents requests must include the sequenceToken obtained from the response of the previous request.

plesNextSequenceToken :: Iso' PutLogEventsResponse (Maybe Text)

data PutMetricFilter

Constructors

PutMetricFilter

Fields

_pmfrFilterName :: Text

The name of the metric filter.

_pmfrFilterPattern :: Text

_pmfrLogGroupName :: Text

_pmfrMetricTransformations :: [MetricTransformation]

pmfrMetricTransformations :: Lens' PutMetricFilter [MetricTransformation]

pmfrLogGroupName :: Lens' PutMetricFilter Text

pmfrFilterPattern :: Lens' PutMetricFilter Text

pmfrFilterName :: Lens' PutMetricFilter Text

data PutMetricFilterResponse

Constructors

PutMetricFilterResponse

data PutRetentionPolicy

Constructors

PutRetentionPolicy

Fields

_prprRetentionInDays :: Integer

Specifies the number of days you want to retain log events in the specified log group. Possible values are: 1, 3, 5, 7, 14, 30, 60, 90, 120, 150, 180, 365, 400, 547, 730.

_prprLogGroupName :: Text

prprRetentionInDays :: Lens' PutRetentionPolicy Integer

prprLogGroupName :: Lens' PutRetentionPolicy Text

data PutRetentionPolicyResponse

Constructors

PutRetentionPolicyResponse

data TestMetricFilter

Constructors

TestMetricFilter

Fields

_tmfrFilterPattern :: Text

_tmfrLogEventMessages :: [Text]

tmfrLogEventMessages :: Lens' TestMetricFilter [Text]

tmfrFilterPattern :: Lens' TestMetricFilter Text

data TestMetricFilterResponse

Constructors

TestMetricFilterResponse

Fields

_tmfsMatches :: [MetricFilterMatchRecord]

tmfsMatches :: Iso' TestMetricFilterResponse [MetricFilterMatchRecord]

data CloudWatchLogs

Supported version (2014-03-28) of the Amazon CloudWatch Logs service.

data InputLogEvent

A log event is a record of some activity that was recorded by the application or resource being monitored. The log event record that Amazon CloudWatch Logs understands contains two properties: the timestamp of when the event occurred, and the raw event message.

Constructors

InputLogEvent

Fields

_ilfMessage :: Text

_ilfTimestamp :: Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

data LogGroup

Constructors

LogGroup

Fields

_lhCreationTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_lhMetricFilterCount :: Maybe Integer

The number of metric filters associated with the log group.

_lhArn :: Maybe Text

_lhLogGroupName :: Maybe Text

_lhRetentionInDays :: Maybe Integer

Specifies the number of days you want to retain log events in the specified log group. Possible values are: 1, 3, 5, 7, 14, 30, 60, 90, 120, 150, 180, 365, 400, 547, 730.

_lhStoredBytes :: Maybe Integer

data LogStream

A log stream is sequence of log events that share the same emitter.

Constructors

LogStream

Fields

_ltCreationTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_ltUploadSequenceToken :: Maybe Text

A string token used for making PutLogEvents requests. A sequenceToken can only be used once, and PutLogEvents requests must include the sequenceToken obtained from the response of the previous request.

_ltArn :: Maybe Text

_ltFirstEventTimestamp :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_ltLogStreamName :: Maybe Text

_ltStoredBytes :: Maybe Integer

_ltLastIngestionTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_ltLastEventTimestamp :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

data MetricFilter

Metric filters can be used to express how Amazon CloudWatch Logs would extract metric observations from ingested log events and transform them to metric data in a CloudWatch metric.

Constructors

MetricFilter

Fields

_mgCreationTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_mgFilterName :: Maybe Text

The name of the metric filter.

_mgFilterPattern :: Maybe Text

_mgMetricTransformations :: Maybe [MetricTransformation]

data MetricFilterMatchRecord

Constructors

MetricFilterMatchRecord

Fields

_mfmrExtractedValues :: Map Text Text

_mfmrEventNumber :: Maybe Integer

_mfmrEventMessage :: Maybe Text

data MetricTransformation

Constructors

MetricTransformation

Fields

_muMetricName :: Text

_muMetricNamespace :: Text

_muMetricValue :: Text

data OutputLogEvent

Constructors

OutputLogEvent

Fields

_olfIngestionTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_olfMessage :: Maybe Text

_olfTimestamp :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

ilfTimestamp :: Lens' InputLogEvent Integer

ilfMessage :: Lens' InputLogEvent Text

lhStoredBytes :: Lens' LogGroup (Maybe Integer)

lhRetentionInDays :: Lens' LogGroup (Maybe Integer)

lhMetricFilterCount :: Lens' LogGroup (Maybe Integer)

lhLogGroupName :: Lens' LogGroup (Maybe Text)

lhCreationTime :: Lens' LogGroup (Maybe Integer)

lhArn :: Lens' LogGroup (Maybe Text)

ltUploadSequenceToken :: Lens' LogStream (Maybe Text)

ltStoredBytes :: Lens' LogStream (Maybe Integer)

ltLogStreamName :: Lens' LogStream (Maybe Text)

ltLastIngestionTime :: Lens' LogStream (Maybe Integer)

ltLastEventTimestamp :: Lens' LogStream (Maybe Integer)

ltFirstEventTimestamp :: Lens' LogStream (Maybe Integer)

ltCreationTime :: Lens' LogStream (Maybe Integer)

ltArn :: Lens' LogStream (Maybe Text)

mgMetricTransformations :: Lens' MetricFilter (Maybe [MetricTransformation])

mgFilterPattern :: Lens' MetricFilter (Maybe Text)

mgFilterName :: Lens' MetricFilter (Maybe Text)

mgCreationTime :: Lens' MetricFilter (Maybe Integer)

mfmrExtractedValues :: Lens' MetricFilterMatchRecord (Map Text Text)

mfmrEventNumber :: Lens' MetricFilterMatchRecord (Maybe Integer)

mfmrEventMessage :: Lens' MetricFilterMatchRecord (Maybe Text)

muMetricValue :: Lens' MetricTransformation Text

muMetricNamespace :: Lens' MetricTransformation Text

muMetricName :: Lens' MetricTransformation Text

olfTimestamp :: Lens' OutputLogEvent (Maybe Integer)

olfMessage :: Lens' OutputLogEvent (Maybe Text)

olfIngestionTime :: Lens' OutputLogEvent (Maybe Integer)