Network.AWS.CloudWatchLogs.V2014_03_28.CreateLogStream

Description

Creates a new log stream in the specified log group. The name of the log stream must be unique within the log group. There is no limit on the number of log streams that can exist in a log group. You must use the following guidelines when naming a log stream: Log stream names can be between 1 and 512 characters long. The : colon character is not allowed. Create a new Log Stream The following is an example of a CreateLogStream request and response. POST HTTP1.1 Host: logs.. X-Amz-Date: Authorization: AWS4-HMAC-SHA256 Credential=, SignedHeaders=content-type;date;host;user-agent;x-amz-date;x-amz-target;x-amzn-requestid, Signature= User-Agent: Accept: application/json Content-Type: application/x-amz-json-1.1 Content-Length: Connection: Keep-Alive]]> X-Amz-Target: Logs_20140328.CreateLogStream { "logGroupName": "exampleLogGroupName", "logStreamName": "exampleLogStreamName" } HTTP/1.1 200 OK x-amzn-RequestId: Content-Type: application/x-amz-json-1.1 Content-Length: Date: ]]>.

Documentation

data CreateLogStream

Constructors

CreateLogStream

Fields

_clsrLogGroupName :: Text

_clsrLogStreamName :: Text

clsrLogStreamName :: Lens' CreateLogStream Text

clsrLogGroupName :: Lens' CreateLogStream Text

data CreateLogStreamResponse

Constructors

CreateLogStreamResponse