Network.AWS.CloudWatchLogs.V2014_03_28.DeleteLogStream

Description

Deletes a log stream and permanently deletes all the archived log events associated with it. Delete a Log Stream The following is an example of a DeleteLogStream request and response. POST HTTP1.1 Host: logs.. X-Amz-Date: Authorization: AWS4-HMAC-SHA256 Credential=, SignedHeaders=content-type;date;host;user-agent;x-amz-date;x-amz-target;x-amzn-requestid, Signature= User-Agent: Accept: application/json Content-Type: application/x-amz-json-1.1 Content-Length: Connection: Keep-Alive]]> X-Amz-Target: Logs_20140328.DeleteLogStream { "logGroupName": "exampleLogGroupName", "logStreamName": "exampleLogStreamName" } HTTP/1.1 200 OK x-amzn-RequestId: Content-Type: application/x-amz-json-1.1 Content-Length: Date: ]]>.

Documentation

data DeleteLogStream

Constructors

DeleteLogStream

Fields

_dlsrLogGroupName :: Text

_dlsrLogStreamName :: Text

dlsrLogStreamName :: Lens' DeleteLogStream Text

dlsrLogGroupName :: Lens' DeleteLogStream Text

data DeleteLogStreamResponse

Constructors

DeleteLogStreamResponse