Network.AWS.CloudWatchLogs.V2014_03_28.DeleteMetricFilter

Description

Deletes a metric filter associated with the specified log group. Delete a metric filter The following is an example of a DeleteMetricFilter request and response. POST HTTP1.1 Host: logs.. X-Amz-Date: Authorization: AWS4-HMAC-SHA256 Credential=, SignedHeaders=content-type;date;host;user-agent;x-amz-date;x-amz-target;x-amzn-requestid, Signature= User-Agent: Accept: application/json Content-Type: application/x-amz-json-1.1 Content-Length: Connection: Keep-Alive]]> X-Amz-Target: Logs_20140328.DeleteMetricFilter { "logGroupName": "exampleLogGroupName", "filterName": "exampleMetricFilterName" } HTTP/1.1 200 OK x-amzn-RequestId: Content-Type: application/x-amz-json-1.1 Content-Length: Date: ]]>.

Request

data DeleteMetricFilter

Request constructor

mkDeleteMetricFilter

Arguments

:: Text
dmfLogGroupName

-> Text
dmfFilterName

-> DeleteMetricFilter

Smart constructor for the minimum required parameters to construct a valid DeleteMetricFilter request.

Request lenses

dmfLogGroupName :: Lens' DeleteMetricFilter Text

dmfFilterName :: Lens' DeleteMetricFilter Text

The name of the metric filter.

Response

data DeleteMetricFilterResponse

Response constructor

mkDeleteMetricFilterResponse :: DeleteMetricFilterResponse

Smart constructor for the minimum required parameters to construct a valid DeleteMetricFilterResponse response.

This constructor is provided for convenience and testing purposes.