Network.AWS.CloudWatchLogs.V2014_03_28.DescribeMetricFilters

Description

Returns all the metrics filters associated with the specified log group. The list returned in the response is ASCII-sorted by filter name. By default, this operation returns up to 50 metric filters. If there are more metric filters to list, the response would contain a nextToken value in the response body. You can also limit the number of metric filters returned in the response by specifying the limit parameter in the request. List the metric filters associated with a log group The following is an example of a DescribeMetricFilters request and response. POST HTTP1.1 Host: logs.. X-Amz-Date: Authorization: AWS4-HMAC-SHA256 Credential=, SignedHeaders=content-type;date;host;user-agent;x-amz-date;x-amz-target;x-amzn-requestid, Signature= User-Agent: Accept: application/json Content-Type: application/x-amz-json-1.1 Content-Length: Connection: Keep-Alive]]> X-Amz-Target: Logs_20140328.DescribeMetricFilters { "logGroupName": "exampleLogGroupName" } HTTP/1.1 200 OK x-amzn-RequestId: Content-Type: application/x-amz-json-1.1 Content-Length: Date: ]]> { "metricFilters": [ { "creationTime": 1396224000000, "filterName": "exampleFilterName", "filterPattern": "[ip, identity, user_id, timestamp, request, status_code, size]", "metricTransformations": [ { "metricValue": "$size", "metricNamespace": MyApp, "metricName": Volume }, { "metricValue": "1", "metricNamespace": MyApp, "metricName": RequestCount } ] } ] }.

Request

data DescribeMetricFilters

Request constructor

mkDescribeMetricFilters

Arguments

:: Text
dmf1LogGroupName

-> DescribeMetricFilters

Smart constructor for the minimum required parameters to construct a valid DescribeMetricFilters request.

The fields accessible through corresponding lenses are:

LogGroupName :: Text

FilterNamePrefix :: Maybe Text

NextToken :: Maybe Text

Limit :: Maybe Integer

Request lenses

dmf1LogGroupName :: Lens' DescribeMetricFilters Text

dmf1FilterNamePrefix :: Lens' DescribeMetricFilters (Maybe Text)

The name of the metric filter.

dmf1NextToken :: Lens' DescribeMetricFilters (Maybe Text)

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous DescribeMetricFilters request.

dmf1Limit :: Lens' DescribeMetricFilters (Maybe Integer)

The maximum number of items returned in the response. If you don't specify a value, the request would return up to 50 items.

Response

data DescribeMetricFiltersResponse

Response constructor

mkDescribeMetricFiltersResponse :: DescribeMetricFiltersResponse

Smart constructor for the minimum required parameters to construct a valid DescribeMetricFiltersResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

MetricFilters :: [MetricFilter]

NextToken :: Maybe Text

Response lenses

dmfrMetricFilters :: Lens' DescribeMetricFiltersResponse [MetricFilter]

dmfrNextToken :: Lens' DescribeMetricFiltersResponse (Maybe Text)

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous request. The token expires after 24 hours.