Network.AWS.CloudWatchLogs.V2014_03_28.GetLogEvents

Description

Retrieves log events from the specified log stream. You can provide an optional time range to filter the results on the event timestamp. By default, this operation returns as much log events as can fit in a response size of 1MB, up to 10,000 log events. The response will always include a nextForwardToken and a nextBackwardToken in the response body. You can use any of these tokens in subsequent GetLogEvents requests to paginate through events in either forward or backward direction. You can also limit the number of log events returned in the response by specifying the limit parameter in the request. Retrieves all the events from a log stream The following is an example of a GetLogEvents request and response. POST / HTTP/1.1 Host: logs.. X-Amz-Date: Authorization: AWS4-HMAC-SHA256 Credential=, SignedHeaders=content-type;date;host;user-agent;x-amz-date;x-amz-target;x-amzn-requestid, Signature= User-Agent: Accept: application/json Content-Type: application/x-amz-json-1.1 Content-Length: Connection: Keep-Alive]]> X-Amz-Target: Logs_20140328.GetLogEvents { "logGroupName": "exampleLogGroupName", "logStreamName": "exampleLogStreamName" } HTTP/1.1 200 OK x-amzn-RequestId: Content-Type: application/x-amz-json-1.1 Content-Length: Date: ]]> { "events": [ { "ingestionTime": 1396035394997, "timestamp": 1396035378988, "message": "Example Event 1" }, { "ingestionTime": 1396035394997, "timestamp": 1396035378988, "message": "Example Event 2" }, { "ingestionTime": 1396035394997, "timestamp": 1396035378989, "message": "Example Event 3" } ], "nextBackwardToken": "b/31132629274945519779805322857203735586714454643391594505", "nextForwardToken": "f/31132629323784151764587387538205132201699397759403884544" }.

Documentation

getLogEvents

Arguments

:: Text
_glerLogGroupName

-> Text
_glerLogStreamName

-> GetLogEvents

Minimum specification for a GetLogEvents request.

data GetLogEvents

Constructors

GetLogEvents

Fields

_glerLogGroupName :: Text

_glerLogStreamName :: Text

_glerLimit :: Maybe Integer

The maximum number of log events returned in the response. If you don't specify a value, the request would return as much log events as can fit in a response size of 1MB, up to 10,000 log events.

_glerNextToken :: Maybe Text

A string token used for pagination that points to the next page of results. It must be a value obtained from the nextForwardToken or nextBackwardToken fields in the response of the previous GetLogEvents request.

_glerStartFromHead :: Maybe Bool

_glerStartTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_glerEndTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

glerStartTime :: Lens' GetLogEvents (Maybe Integer)

glerStartFromHead :: Lens' GetLogEvents (Maybe Bool)

glerNextToken :: Lens' GetLogEvents (Maybe Text)

glerLogStreamName :: Lens' GetLogEvents Text

glerLogGroupName :: Lens' GetLogEvents Text

glerLimit :: Lens' GetLogEvents (Maybe Integer)

glerEndTime :: Lens' GetLogEvents (Maybe Integer)

data GetLogEventsResponse

Constructors

GetLogEventsResponse

Fields

_glesNextBackwardToken :: Maybe Text

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous request. The token expires after 24 hours.

_glesNextForwardToken :: Maybe Text

A string token used for pagination that points to the next page of results. It must be a value obtained from the response of the previous request. The token expires after 24 hours.

_glesEvents :: [OutputLogEvent]

glesNextForwardToken :: Lens' GetLogEventsResponse (Maybe Text)

glesNextBackwardToken :: Lens' GetLogEventsResponse (Maybe Text)

glesEvents :: Lens' GetLogEventsResponse [OutputLogEvent]