Network.AWS.CloudWatchLogs.V2014_03_28.PutMetricFilter

Description

Creates or updates a metric filter and associates it with the specified log group. Metric filters allow you to configure rules to extract metric data from log events ingested through PutLogEvents requests. Create or update a metric filter The following is an example of a PutMetricFilter request and response. POST HTTP1.1 Host: logs.. X-Amz-Date: Authorization: AWS4-HMAC-SHA256 Credential=, SignedHeaders=content-type;date;host;user-agent;x-amz-date;x-amz-target;x-amzn-requestid, Signature= User-Agent: Accept: application/json Content-Type: application/x-amz-json-1.1 Content-Length: Connection: Keep-Alive]]> X-Amz-Target: Logs_20140328.PutMetricFilter { "logGroupName": "exampleLogGroupName", "filterName": "exampleMetricFilterName", "filterPattern": "[ip, identity, user_id, timestamp, request, status_code, size]", "metricTransformations": [ { "metricValue": "$size", "metricNamespace": MyApp, "metricName": Volume }, { "metricValue": "1", "metricNamespace": MyApp, "metricName": RequestCount } ] } HTTP/1.1 200 OK x-amzn-RequestId: Content-Type: application/x-amz-json-1.1 Content-Length: Date: ]]>.

Request

data PutMetricFilter

Request constructor

mkPutMetricFilter

Arguments

:: Text
pmfLogGroupName

-> Text
pmfFilterName

-> Text
pmfFilterPattern

-> List1 MetricTransformation
pmfMetricTransformations

-> PutMetricFilter

Smart constructor for the minimum required parameters to construct a valid PutMetricFilter request.

The fields accessible through corresponding lenses are:

LogGroupName :: Text

FilterName :: Text

FilterPattern :: Text

MetricTransformations :: List1 MetricTransformation

Request lenses

pmfLogGroupName :: Lens' PutMetricFilter Text

pmfFilterName :: Lens' PutMetricFilter Text

The name of the metric filter.

pmfFilterPattern :: Lens' PutMetricFilter Text

pmfMetricTransformations :: Lens' PutMetricFilter (List1 MetricTransformation)

Response

data PutMetricFilterResponse

Response constructor

mkPutMetricFilterResponse :: PutMetricFilterResponse

Smart constructor for the minimum required parameters to construct a valid PutMetricFilterResponse response.

This constructor is provided for convenience and testing purposes.