Network.AWS.CloudWatchLogs.V2014_03_28.Types

Description

Amazon CloudWatch Logs enables you to monitor, store, and access your system, application, and custom log files.

Service

data CloudWatchLogs

Supported version (2014-03-28) of the Amazon CloudWatch Logs service.

InputLogEvent

data InputLogEvent

A log event is a record of some activity that was recorded by the application or resource being monitored. The log event record that Amazon CloudWatch Logs understands contains two properties: the timestamp of when the event occurred, and the raw event message.

mkInputLogEvent

Arguments

:: Integer
ileTimestamp

-> Text
ileMessage

-> InputLogEvent

Smart constructor for the minimum required fields to construct a valid InputLogEvent data type to populate a request.

ileTimestamp :: Lens' InputLogEvent Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

ileMessage :: Lens' InputLogEvent Text

LogGroup

data LogGroup

mkLogGroup :: LogGroup

Smart constructor for the minimum required fields to construct a valid LogGroup data type to populate a request.

lgLogGroupName :: Lens' LogGroup (Maybe Text)

lgCreationTime :: Lens' LogGroup (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

lgRetentionInDays :: Lens' LogGroup (Maybe Integer)

Specifies the number of days you want to retain log events in the specified log group. Possible values are: 1, 3, 5, 7, 14, 30, 60, 90, 120, 150, 180, 365, 400, 547, 730.

lgMetricFilterCount :: Lens' LogGroup (Maybe Integer)

The number of metric filters associated with the log group.

lgArn :: Lens' LogGroup (Maybe Text)

lgStoredBytes :: Lens' LogGroup (Maybe Integer)

LogStream

data LogStream

A log stream is sequence of log events that share the same emitter.

mkLogStream :: LogStream

Smart constructor for the minimum required fields to construct a valid LogStream data type to populate a request.

lsLogStreamName :: Lens' LogStream (Maybe Text)

lsCreationTime :: Lens' LogStream (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

lsFirstEventTimestamp :: Lens' LogStream (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

lsLastEventTimestamp :: Lens' LogStream (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

lsLastIngestionTime :: Lens' LogStream (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

lsUploadSequenceToken :: Lens' LogStream (Maybe Text)

A string token used for making PutLogEvents requests. A sequenceToken can only be used once, and PutLogEvents requests must include the sequenceToken obtained from the response of the previous request.

lsArn :: Lens' LogStream (Maybe Text)

lsStoredBytes :: Lens' LogStream (Maybe Integer)

MetricFilter

data MetricFilter

Metric filters can be used to express how Amazon CloudWatch Logs would extract metric observations from ingested log events and transform them to metric data in a CloudWatch metric.

mkMetricFilter :: MetricFilter

Smart constructor for the minimum required fields to construct a valid MetricFilter data type to populate a request.

mfFilterName :: Lens' MetricFilter (Maybe Text)

The name of the metric filter.

mfFilterPattern :: Lens' MetricFilter (Maybe Text)

mfMetricTransformations :: Lens' MetricFilter (Maybe (List1 MetricTransformation))

mfCreationTime :: Lens' MetricFilter (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

MetricFilterMatchRecord

data MetricFilterMatchRecord

mkMetricFilterMatchRecord :: MetricFilterMatchRecord

Smart constructor for the minimum required fields to construct a valid MetricFilterMatchRecord data type to populate a request.

mfmrEventNumber :: Lens' MetricFilterMatchRecord (Maybe Integer)

mfmrEventMessage :: Lens' MetricFilterMatchRecord (Maybe Text)

mfmrExtractedValues :: Lens' MetricFilterMatchRecord (Map Text Text)

MetricTransformation

data MetricTransformation

mkMetricTransformation

Arguments

:: Text
mtMetricName

-> Text
mtMetricNamespace

-> Text
mtMetricValue

-> MetricTransformation

Smart constructor for the minimum required fields to construct a valid MetricTransformation data type to populate a request.

mtMetricName :: Lens' MetricTransformation Text

mtMetricNamespace :: Lens' MetricTransformation Text

mtMetricValue :: Lens' MetricTransformation Text

OutputLogEvent

data OutputLogEvent

mkOutputLogEvent :: OutputLogEvent

Smart constructor for the minimum required fields to construct a valid OutputLogEvent data type to populate a request.

oleTimestamp :: Lens' OutputLogEvent (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

oleMessage :: Lens' OutputLogEvent (Maybe Text)

oleIngestionTime :: Lens' OutputLogEvent (Maybe Integer)

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.