Network.AWS.CloudWatchLogs.DescribeLogGroups

Description

Lists the specified log groups. You can list all your log groups or filter the results by prefix. The results are ASCII-sorted by log group name.

This operation returns paginated results.

Creating a Request

describeLogGroups :: DescribeLogGroups

Creates a value of DescribeLogGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlgLogGroupNamePrefix - The prefix to match.

dlgNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dlgLimit - The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

data DescribeLogGroups

See: describeLogGroups smart constructor.

Request Lenses

dlgLogGroupNamePrefix :: Lens' DescribeLogGroups (Maybe Text)

The prefix to match.

dlgNextToken :: Lens' DescribeLogGroups (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dlgLimit :: Lens' DescribeLogGroups (Maybe Natural)

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

Destructuring the Response

describeLogGroupsResponse

Arguments

:: Int
dlgrsResponseStatus

-> DescribeLogGroupsResponse

Creates a value of DescribeLogGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlgrsLogGroups - The log groups.

dlgrsNextToken - Undocumented member.

dlgrsResponseStatus - -- | The response status code.

data DescribeLogGroupsResponse

See: describeLogGroupsResponse smart constructor.

Response Lenses

dlgrsLogGroups :: Lens' DescribeLogGroupsResponse [LogGroup]

The log groups.

dlgrsNextToken :: Lens' DescribeLogGroupsResponse (Maybe Text)

Undocumented member.

dlgrsResponseStatus :: Lens' DescribeLogGroupsResponse Int

- | The response status code.