Network.AWS.CloudWatchLogs.DescribeLogStreams

Description

Lists the log streams for the specified log group. You can list all the log streams or filter the results by prefix. You can also control how the results are ordered.

This operation has a limit of five transactions per second, after which transactions are throttled.

This operation returns paginated results.

Creating a Request

describeLogStreams

Arguments

:: Text
dlssLogGroupName

-> DescribeLogStreams

Creates a value of DescribeLogStreams with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlssOrderBy - If the value is LogStreamName , the results are ordered by log stream name. If the value is LastEventTime , the results are ordered by the event time. The default value is LogStreamName . If you order the results by event time, you cannot specify the logStreamNamePrefix parameter. lastEventTimestamp represents the time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. lastEventTimeStamp updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but may take longer in some rare situations.

dlssDescending - If the value is true, results are returned in descending order. If the value is to false, results are returned in ascending order. The default value is false.

dlssNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dlssLogStreamNamePrefix - The prefix to match. You cannot specify this parameter if orderBy is LastEventTime .

dlssLimit - The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

dlssLogGroupName - The name of the log group.

data DescribeLogStreams

See: describeLogStreams smart constructor.

Request Lenses

dlssOrderBy :: Lens' DescribeLogStreams (Maybe OrderBy)

If the value is LogStreamName , the results are ordered by log stream name. If the value is LastEventTime , the results are ordered by the event time. The default value is LogStreamName . If you order the results by event time, you cannot specify the logStreamNamePrefix parameter. lastEventTimestamp represents the time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. lastEventTimeStamp updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but may take longer in some rare situations.

dlssDescending :: Lens' DescribeLogStreams (Maybe Bool)

If the value is true, results are returned in descending order. If the value is to false, results are returned in ascending order. The default value is false.

dlssNextToken :: Lens' DescribeLogStreams (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dlssLogStreamNamePrefix :: Lens' DescribeLogStreams (Maybe Text)

The prefix to match. You cannot specify this parameter if orderBy is LastEventTime .

dlssLimit :: Lens' DescribeLogStreams (Maybe Natural)

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

dlssLogGroupName :: Lens' DescribeLogStreams Text

The name of the log group.

Destructuring the Response

describeLogStreamsResponse

Arguments

:: Int
dlsrsResponseStatus

-> DescribeLogStreamsResponse

Creates a value of DescribeLogStreamsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlsrsNextToken - Undocumented member.

dlsrsLogStreams - The log streams.

dlsrsResponseStatus - -- | The response status code.

data DescribeLogStreamsResponse

See: describeLogStreamsResponse smart constructor.

Response Lenses

dlsrsNextToken :: Lens' DescribeLogStreamsResponse (Maybe Text)

Undocumented member.

dlsrsLogStreams :: Lens' DescribeLogStreamsResponse [LogStream]

The log streams.

dlsrsResponseStatus :: Lens' DescribeLogStreamsResponse Int

- | The response status code.