Network.AWS.CloudWatchLogs.DescribeMetricFilters

Description

Lists the specified metric filters. You can list all the metric filters or filter the results by log name, prefix, metric name, and metric namespace. The results are ASCII-sorted by filter name.

This operation returns paginated results.

Creating a Request

describeMetricFilters :: DescribeMetricFilters

Creates a value of DescribeMetricFilters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmfFilterNamePrefix - The prefix to match.

dmfMetricName - The name of the CloudWatch metric.

dmfLogGroupName - The name of the log group.

dmfNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dmfMetricNamespace - The namespace of the CloudWatch metric.

dmfLimit - The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

data DescribeMetricFilters

See: describeMetricFilters smart constructor.

Request Lenses

dmfFilterNamePrefix :: Lens' DescribeMetricFilters (Maybe Text)

The prefix to match.

dmfMetricName :: Lens' DescribeMetricFilters (Maybe Text)

The name of the CloudWatch metric.

dmfLogGroupName :: Lens' DescribeMetricFilters (Maybe Text)

The name of the log group.

dmfNextToken :: Lens' DescribeMetricFilters (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dmfMetricNamespace :: Lens' DescribeMetricFilters (Maybe Text)

The namespace of the CloudWatch metric.

dmfLimit :: Lens' DescribeMetricFilters (Maybe Natural)

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

Destructuring the Response

describeMetricFiltersResponse

Arguments

:: Int
dmfrsResponseStatus

-> DescribeMetricFiltersResponse

Creates a value of DescribeMetricFiltersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmfrsNextToken - Undocumented member.

dmfrsMetricFilters - The metric filters.

dmfrsResponseStatus - -- | The response status code.

data DescribeMetricFiltersResponse

See: describeMetricFiltersResponse smart constructor.

Response Lenses

dmfrsNextToken :: Lens' DescribeMetricFiltersResponse (Maybe Text)

Undocumented member.

dmfrsMetricFilters :: Lens' DescribeMetricFiltersResponse [MetricFilter]

The metric filters.

dmfrsResponseStatus :: Lens' DescribeMetricFiltersResponse Int

- | The response status code.