Network.AWS.CloudWatchLogs.GetLogEvents

Description

Lists log events from the specified log stream. You can list all the log events or filter using a time range.

By default, this operation returns as many log events as can fit in a response size of 1MB (up to 10,000 log events). If the results include tokens, there are more log events available. You can get additional log events by specifying one of the tokens in a subsequent call.

Creating a Request

getLogEvents

Arguments

:: Text
gleLogGroupName

-> Text
gleLogStreamName

-> GetLogEvents

Creates a value of GetLogEvents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gleStartTime - The start of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp earlier than this time are not included.

gleStartFromHead - If the value is true, the earliest log events are returned first. If the value is false, the latest log events are returned first. The default value is false.

gleNextToken - The token for the next set of items to return. (You received this token from a previous call.)

gleEndTime - The end of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not included.

gleLimit - The maximum number of log events returned. If you don't specify a value, the maximum is as many log events as can fit in a response size of 1MB, up to 10,000 log events.

gleLogGroupName - The name of the log group.

gleLogStreamName - The name of the log stream.

data GetLogEvents

See: getLogEvents smart constructor.

Request Lenses

gleStartTime :: Lens' GetLogEvents (Maybe Natural)

The start of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp earlier than this time are not included.

gleStartFromHead :: Lens' GetLogEvents (Maybe Bool)

If the value is true, the earliest log events are returned first. If the value is false, the latest log events are returned first. The default value is false.

gleNextToken :: Lens' GetLogEvents (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

gleEndTime :: Lens' GetLogEvents (Maybe Natural)

The end of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not included.

gleLimit :: Lens' GetLogEvents (Maybe Natural)

The maximum number of log events returned. If you don't specify a value, the maximum is as many log events as can fit in a response size of 1MB, up to 10,000 log events.

gleLogGroupName :: Lens' GetLogEvents Text

The name of the log group.

gleLogStreamName :: Lens' GetLogEvents Text

The name of the log stream.

Destructuring the Response

getLogEventsResponse

Arguments

:: Int
glersResponseStatus

-> GetLogEventsResponse

Creates a value of GetLogEventsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

glersNextBackwardToken - The token for the next set of items in the backward direction. The token expires after 24 hours.

glersNextForwardToken - The token for the next set of items in the forward direction. The token expires after 24 hours.

glersEvents - The events.

glersResponseStatus - -- | The response status code.

data GetLogEventsResponse

See: getLogEventsResponse smart constructor.

Response Lenses

glersNextBackwardToken :: Lens' GetLogEventsResponse (Maybe Text)

The token for the next set of items in the backward direction. The token expires after 24 hours.

glersNextForwardToken :: Lens' GetLogEventsResponse (Maybe Text)

The token for the next set of items in the forward direction. The token expires after 24 hours.

glersEvents :: Lens' GetLogEventsResponse [OutputLogEvent]

The events.

glersResponseStatus :: Lens' GetLogEventsResponse Int

- | The response status code.