Network.AWS.CloudWatchLogs.GetLogEvents

Description

Retrieves log events from the specified log stream. You can provide an optional time range to filter the results on the event timestamp.

By default, this operation returns as much log events as can fit in a response size of 1MB, up to 10,000 log events. The response will always include a nextForwardToken and a nextBackwardToken in the response body. You can use any of these tokens in subsequent GetLogEvents requests to paginate through events in either forward or backward direction. You can also limit the number of log events returned in the response by specifying the limit parameter in the request.

http://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_GetLogEvents.html

Request

data GetLogEvents

Request constructor

getLogEvents

Arguments

:: Text
gleLogGroupName

-> Text
gleLogStreamName

-> GetLogEvents

GetLogEvents constructor.

The fields accessible through corresponding lenses are:

gleEndTime :: Maybe Natural

gleLimit :: Maybe Natural

gleLogGroupName :: Text

gleLogStreamName :: Text

gleNextToken :: Maybe Text

gleStartFromHead :: Maybe Bool

gleStartTime :: Maybe Natural

Request lenses

gleEndTime :: Lens' GetLogEvents (Maybe Natural)

gleLimit :: Lens' GetLogEvents (Maybe Natural)

The maximum number of log events returned in the response. If you don't specify a value, the request would return as much log events as can fit in a response size of 1MB, up to 10,000 log events.

gleLogGroupName :: Lens' GetLogEvents Text

gleLogStreamName :: Lens' GetLogEvents Text

gleNextToken :: Lens' GetLogEvents (Maybe Text)

A string token used for pagination that points to the next page of results. It must be a value obtained from the nextForwardToken or nextBackwardToken fields in the response of the previous GetLogEvents request.

gleStartFromHead :: Lens' GetLogEvents (Maybe Bool)

If set to true, the earliest log events would be returned first. The default is false (the latest log events are returned first).

gleStartTime :: Lens' GetLogEvents (Maybe Natural)

Response

data GetLogEventsResponse

Response constructor

getLogEventsResponse :: GetLogEventsResponse

GetLogEventsResponse constructor.

The fields accessible through corresponding lenses are:

glerEvents :: [OutputLogEvent]

glerNextBackwardToken :: Maybe Text

glerNextForwardToken :: Maybe Text

Response lenses

glerEvents :: Lens' GetLogEventsResponse [OutputLogEvent]

glerNextBackwardToken :: Lens' GetLogEventsResponse (Maybe Text)

glerNextForwardToken :: Lens' GetLogEventsResponse (Maybe Text)