Network.AWS.CloudWatchLogs.PutLogEvents

Description

Uploads a batch of log events to the specified log stream.

You must include the sequence token obtained from the response of the previous call. An upload in a newly created log stream does not require a sequence token. You can also get the sequence token using DescribeLogStreams .

The batch of events must satisfy the following constraints:

The maximum batch size is 1,048,576 bytes, and this size is calculated as the sum of all event messages in UTF-8, plus 26 bytes for each log event.

None of the log events in the batch can be more than 2 hours in the future.

None of the log events in the batch can be older than 14 days or the retention period of the log group.

The log events in the batch must be in chronological ordered by their timestamp (the time the event occurred, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC).

The maximum number of log events in a batch is 10,000.

A batch of log events in a single request cannot span more than 24 hours. Otherwise, the operation fails.

Creating a Request

putLogEvents

Arguments

:: Text
pleLogGroupName

-> Text
pleLogStreamName

-> NonEmpty InputLogEvent
pleLogEvents

-> PutLogEvents

Creates a value of PutLogEvents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pleSequenceToken - The sequence token.

pleLogGroupName - The name of the log group.

pleLogStreamName - The name of the log stream.

pleLogEvents - The log events.

data PutLogEvents

See: putLogEvents smart constructor.

Request Lenses

pleSequenceToken :: Lens' PutLogEvents (Maybe Text)

The sequence token.

pleLogGroupName :: Lens' PutLogEvents Text

The name of the log group.

pleLogStreamName :: Lens' PutLogEvents Text

The name of the log stream.

pleLogEvents :: Lens' PutLogEvents (NonEmpty InputLogEvent)

The log events.

Destructuring the Response

putLogEventsResponse

Arguments

:: Int
plersResponseStatus

-> PutLogEventsResponse

Creates a value of PutLogEventsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

plersRejectedLogEventsInfo - The rejected events.

plersNextSequenceToken - The next sequence token.

plersResponseStatus - -- | The response status code.

data PutLogEventsResponse

See: putLogEventsResponse smart constructor.

Response Lenses

plersRejectedLogEventsInfo :: Lens' PutLogEventsResponse (Maybe RejectedLogEventsInfo)

The rejected events.

plersNextSequenceToken :: Lens' PutLogEventsResponse (Maybe Text)

The next sequence token.

plersResponseStatus :: Lens' PutLogEventsResponse Int

- | The response status code.