Network.AWS.CloudWatchLogs.PutLogEvents

Description

Uploads a batch of log events to the specified log stream. Every PutLogEvents request must include the sequenceToken obtained from the response of the previous request. An upload in a newly created log stream does not require a sequenceToken. The batch of events must satisfy the following constraints: The maximum batch size is 32,768 bytes, and this size is calculated as the sum of all event messages in UTF-8, plus 26 bytes for each log event. None of the log events in the batch can be more than 2 hours in the future. None of the log events in the batch can be older than 14 days or the retention period of the log group. The log events in the batch must be in chronological ordered by their timestamp. The maximum number of log events in a batch is 1,000. Upload a batch of log events into a log stream The following is an example of a PutLogEvents request and response. POST HTTP1.1 Host: logs.. X-Amz-Date: Authorization: AWS4-HMAC-SHA256 Credential=, SignedHeaders=content-type;date;host;user-agent;x-amz-date;x-amz-target;x-amzn-requestid, Signature= User-Agent: Accept: application/json Content-Type: application/x-amz-json-1.1 Content-Length: Connection: Keep-Alive]]> X-Amz-Target: Logs_20140328.PutLogEvents { "logGroupName": "exampleLogGroupName", "logStreamName": "exampleLogStreamName", "logEvents": [ { "timestamp": 1396035378988, "message": "Example Event 1" }, { "timestamp": 1396035378988, "message": "Example Event 2" }, { "timestamp": 1396035378989, "message": "Example Event 3" } ] } HTTP/1.1 200 OK x-amzn-RequestId: Content-Type: application/x-amz-json-1.1 Content-Length: Date: ]]> { "nextSequenceToken": "49536701251539826331025683274032969384950891766572122113" }.

Request

data PutLogEvents

Request constructor

mkPutLogEvents

Arguments

:: Text
pleLogGroupName

-> Text
pleLogStreamName

-> List1 InputLogEvent
pleLogEvents

-> PutLogEvents

Smart constructor for the minimum required parameters to construct a valid PutLogEvents request.

The fields accessible through corresponding lenses are:

LogGroupName :: Text

LogStreamName :: Text

LogEvents :: List1 InputLogEvent

SequenceToken :: Maybe Text

Request lenses

pleLogGroupName :: Lens' PutLogEvents Text

pleLogStreamName :: Lens' PutLogEvents Text

pleLogEvents :: Lens' PutLogEvents (List1 InputLogEvent)

A list of events belonging to a log stream.

pleSequenceToken :: Lens' PutLogEvents (Maybe Text)

A string token that must be obtained from the response of the previous PutLogEvents request.

Response

data PutLogEventsResponse

Response constructor

mkPutLogEventsResponse :: PutLogEventsResponse

Smart constructor for the minimum required parameters to construct a valid PutLogEventsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

NextSequenceToken :: Maybe Text

Response lenses

plerNextSequenceToken :: Lens' PutLogEventsResponse (Maybe Text)

A string token used for making PutLogEvents requests. A sequenceToken can only be used once, and PutLogEvents requests must include the sequenceToken obtained from the response of the previous request.