Network.AWS.CloudWatchLogs.PutMetricFilter

Description

Creates or updates a metric filter and associates it with the specified log group. Metric filters allow you to configure rules to extract metric data from log events ingested through PutLogEvents .

The maximum number of metric filters that can be associated with a log group is 100.

Creating a Request

putMetricFilter

Arguments

:: Text
pmfLogGroupName

-> Text
pmfFilterName

-> Text
pmfFilterPattern

-> NonEmpty MetricTransformation
pmfMetricTransformations

-> PutMetricFilter

Creates a value of PutMetricFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pmfLogGroupName - The name of the log group.

pmfFilterName - A name for the metric filter.

pmfFilterPattern - A filter pattern for extracting metric data out of ingested log events.

pmfMetricTransformations - A collection of information needed to define how metric data gets emitted.

data PutMetricFilter

See: putMetricFilter smart constructor.

Request Lenses

pmfLogGroupName :: Lens' PutMetricFilter Text

The name of the log group.

pmfFilterName :: Lens' PutMetricFilter Text

A name for the metric filter.

pmfFilterPattern :: Lens' PutMetricFilter Text

A filter pattern for extracting metric data out of ingested log events.

pmfMetricTransformations :: Lens' PutMetricFilter (NonEmpty MetricTransformation)

A collection of information needed to define how metric data gets emitted.

Destructuring the Response

putMetricFilterResponse :: PutMetricFilterResponse

Creates a value of PutMetricFilterResponse with the minimum fields required to make a request.

data PutMetricFilterResponse

See: putMetricFilterResponse smart constructor.