Network.AWS.CloudWatchLogs.PutRetentionPolicy

Description

Sets the retention of the specified log group. A retention policy allows you to configure the number of days you want to retain log events in the specified log group.

http://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_PutRetentionPolicy.html

Request

data PutRetentionPolicy

Request constructor

putRetentionPolicy

Arguments

:: Text
prpLogGroupName

-> Int
prpRetentionInDays

-> PutRetentionPolicy

PutRetentionPolicy constructor.

The fields accessible through corresponding lenses are:

prpLogGroupName :: Text

prpRetentionInDays :: Int

Request lenses

prpLogGroupName :: Lens' PutRetentionPolicy Text

prpRetentionInDays :: Lens' PutRetentionPolicy Int

Response

data PutRetentionPolicyResponse

Response constructor

putRetentionPolicyResponse :: PutRetentionPolicyResponse

PutRetentionPolicyResponse constructor.