Network.AWS.CloudWatchLogs.TestMetricFilter

Description

Tests the filter pattern of a metric filter against a sample of log event messages. You can use this operation to validate the correctness of a metric filter pattern.

http://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_TestMetricFilter.html

Request

data TestMetricFilter

Request constructor

testMetricFilter

Arguments

:: Text
tmfFilterPattern

-> NonEmpty Text
tmfLogEventMessages

-> TestMetricFilter

TestMetricFilter constructor.

The fields accessible through corresponding lenses are:

tmfFilterPattern :: Text

tmfLogEventMessages :: NonEmpty Text

Request lenses

tmfFilterPattern :: Lens' TestMetricFilter Text

tmfLogEventMessages :: Lens' TestMetricFilter (NonEmpty Text)

Response

data TestMetricFilterResponse

Response constructor

testMetricFilterResponse :: TestMetricFilterResponse

TestMetricFilterResponse constructor.

The fields accessible through corresponding lenses are:

tmfrMatches :: [MetricFilterMatchRecord]

Response lenses

tmfrMatches :: Lens' TestMetricFilterResponse [MetricFilterMatchRecord]