Network.AWS.CloudWatchLogs.V2014_03_28.DeleteLogGroup

Description

Deletes the log group with the specified name and permanently deletes all the archived log events associated with it. Delete a Log Group The following is an example of a DeleteLogGroup request and response. POST / HTTP/1.1 Host: logs.. X-Amz-Date: Authorization: AWS4-HMAC-SHA256 Credential=, SignedHeaders=content-type;date;host;user-agent;x-amz-date;x-amz-target;x-amzn-requestid, Signature= User-Agent: Accept: application/json Content-Type: application/x-amz-json-1.1 Content-Length: Connection: Keep-Alive]]> X-Amz-Target: Logs_20140328.DeleteLogGroup { "logGroupName": "exampleLogGroupName" } HTTP/1.1 200 OK x-amzn-RequestId: Content-Type: application/x-amz-json-1.1 Content-Length: Date: ]]>.

Request

data DeleteLogGroup

Request constructor

mkDeleteLogGroup

Arguments

:: Text
dlgLogGroupName

-> DeleteLogGroup

Smart constructor for the minimum required parameters to construct a valid DeleteLogGroup request.

Request lenses

dlgLogGroupName :: Lens' DeleteLogGroup Text

Response

data DeleteLogGroupResponse

Response constructor

mkDeleteLogGroupResponse :: DeleteLogGroupResponse

Smart constructor for the minimum required parameters to construct a valid DeleteLogGroupResponse response.

This constructor is provided for convenience and testing purposes.