Network.AWS.CloudWatchLogs.V2014_03_28.PutMetricFilter

Description

Creates or updates a metric filter and associates it with the specified log group. Metric filters allow you to configure rules to extract metric data from log events ingested through PutLogEvents requests. Create or update a metric filter The following is an example of a PutMetricFilter request and response. POST HTTP1.1 Host: logs.. X-Amz-Date: Authorization: AWS4-HMAC-SHA256 Credential=, SignedHeaders=content-type;date;host;user-agent;x-amz-date;x-amz-target;x-amzn-requestid, Signature= User-Agent: Accept: application/json Content-Type: application/x-amz-json-1.1 Content-Length: Connection: Keep-Alive]]> X-Amz-Target: Logs_20140328.PutMetricFilter { "logGroupName": "exampleLogGroupName", "filterName": "exampleMetricFilterName", "filterPattern": "[ip, identity, user_id, timestamp, request, status_code, size]", "metricTransformations": [ { "metricValue": "$size", "metricNamespace": MyApp, "metricName": Volume }, { "metricValue": "1", "metricNamespace": MyApp, "metricName": RequestCount } ] } HTTP/1.1 200 OK x-amzn-RequestId: Content-Type: application/x-amz-json-1.1 Content-Length: Date: ]]>.

Documentation

data PutMetricFilter

Constructors

PutMetricFilter

Fields

_pmfrFilterName :: Text

The name of the metric filter.

_pmfrFilterPattern :: Text

_pmfrLogGroupName :: Text

_pmfrMetricTransformations :: [MetricTransformation]

pmfrMetricTransformations :: Lens' PutMetricFilter [MetricTransformation]

pmfrLogGroupName :: Lens' PutMetricFilter Text

pmfrFilterPattern :: Lens' PutMetricFilter Text

pmfrFilterName :: Lens' PutMetricFilter Text

data PutMetricFilterResponse

Constructors

PutMetricFilterResponse