Network.AWS.CloudWatchLogs.V2014_03_28.Types

Description

Amazon CloudWatch Logs enables you to monitor, store, and access your system, application, and custom log files.

Documentation

data CloudWatchLogs

Supported version (2014-03-28) of the Amazon CloudWatch Logs service.

data InputLogEvent

A log event is a record of some activity that was recorded by the application or resource being monitored. The log event record that Amazon CloudWatch Logs understands contains two properties: the timestamp of when the event occurred, and the raw event message.

Constructors

InputLogEvent

Fields

_ilfMessage :: Text

_ilfTimestamp :: Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

data LogGroup

Constructors

LogGroup

Fields

_lhCreationTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_lhMetricFilterCount :: Maybe Integer

The number of metric filters associated with the log group.

_lhArn :: Maybe Text

_lhLogGroupName :: Maybe Text

_lhRetentionInDays :: Maybe Integer

Specifies the number of days you want to retain log events in the specified log group. Possible values are: 1, 3, 5, 7, 14, 30, 60, 90, 120, 150, 180, 365, 400, 547, 730.

_lhStoredBytes :: Maybe Integer

data LogStream

A log stream is sequence of log events that share the same emitter.

Constructors

LogStream

Fields

_ltCreationTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_ltUploadSequenceToken :: Maybe Text

A string token used for making PutLogEvents requests. A sequenceToken can only be used once, and PutLogEvents requests must include the sequenceToken obtained from the response of the previous request.

_ltArn :: Maybe Text

_ltFirstEventTimestamp :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_ltLogStreamName :: Maybe Text

_ltStoredBytes :: Maybe Integer

_ltLastIngestionTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_ltLastEventTimestamp :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

data MetricFilter

Metric filters can be used to express how Amazon CloudWatch Logs would extract metric observations from ingested log events and transform them to metric data in a CloudWatch metric.

Constructors

MetricFilter

Fields

_mgCreationTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_mgFilterName :: Maybe Text

The name of the metric filter.

_mgFilterPattern :: Maybe Text

_mgMetricTransformations :: Maybe [MetricTransformation]

data MetricFilterMatchRecord

Constructors

MetricFilterMatchRecord

Fields

_mfmrExtractedValues :: Map Text Text

_mfmrEventNumber :: Maybe Integer

_mfmrEventMessage :: Maybe Text

data MetricTransformation

Constructors

MetricTransformation

Fields

_muMetricName :: Text

_muMetricNamespace :: Text

_muMetricValue :: Text

data OutputLogEvent

Constructors

OutputLogEvent

Fields

_olfIngestionTime :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

_olfMessage :: Maybe Text

_olfTimestamp :: Maybe Integer

A point in time expressed as the number milliseconds since Jan 1, 1970 00:00:00 UTC.

ilfTimestamp :: Lens' InputLogEvent Integer

ilfMessage :: Lens' InputLogEvent Text

lhStoredBytes :: Lens' LogGroup (Maybe Integer)

lhRetentionInDays :: Lens' LogGroup (Maybe Integer)

lhMetricFilterCount :: Lens' LogGroup (Maybe Integer)

lhLogGroupName :: Lens' LogGroup (Maybe Text)

lhCreationTime :: Lens' LogGroup (Maybe Integer)

lhArn :: Lens' LogGroup (Maybe Text)

ltUploadSequenceToken :: Lens' LogStream (Maybe Text)

ltStoredBytes :: Lens' LogStream (Maybe Integer)

ltLogStreamName :: Lens' LogStream (Maybe Text)

ltLastIngestionTime :: Lens' LogStream (Maybe Integer)

ltLastEventTimestamp :: Lens' LogStream (Maybe Integer)

ltFirstEventTimestamp :: Lens' LogStream (Maybe Integer)

ltCreationTime :: Lens' LogStream (Maybe Integer)

ltArn :: Lens' LogStream (Maybe Text)

mgMetricTransformations :: Lens' MetricFilter (Maybe [MetricTransformation])

mgFilterPattern :: Lens' MetricFilter (Maybe Text)

mgFilterName :: Lens' MetricFilter (Maybe Text)

mgCreationTime :: Lens' MetricFilter (Maybe Integer)

mfmrExtractedValues :: Lens' MetricFilterMatchRecord (Map Text Text)

mfmrEventNumber :: Lens' MetricFilterMatchRecord (Maybe Integer)

mfmrEventMessage :: Lens' MetricFilterMatchRecord (Maybe Text)

muMetricValue :: Lens' MetricTransformation Text

muMetricNamespace :: Lens' MetricTransformation Text

muMetricName :: Lens' MetricTransformation Text

olfTimestamp :: Lens' OutputLogEvent (Maybe Integer)

olfMessage :: Lens' OutputLogEvent (Maybe Text)

olfIngestionTime :: Lens' OutputLogEvent (Maybe Integer)